Refactor: Reduce editor-canvas.component.ts from 5900 to 1081 lines
