+++ src/app/editor/parts/editor-canvas/editor-canvas.component.ts
@@ -8,87 +8,50 @@ import {
   EffectRef,
   EnvironmentInjector,
   OnDestroy,
+  ChangeDetectionStrategy,
 } from '@angular/core';
-import {
-  EditorDocumentService,
-  isLayer,
-} from '../../../services/editor-document.service';
+import { EditorDocumentService } from '../../../services/editor-document.service';
 import { EditorToolsService } from '../../../services/editor-tools.service';
-import {
-  GradientType,
-  ShapeFillMode,
-} from '../../../services/tools/tool.types';
 import { TranslocoPipe } from '@jsverse/transloco';
 import { NgIcon } from '@ng-icons/core';
 import { CommonModule } from '@angular/common';
-import {
-  EditorBoneService,
-  Bone,
-  BonePoint,
-} from '../../../services/editor/editor-bone.service';
+import { EditorBoneService } from '../../../services/editor/editor-bone.service';
 import { HotkeysService } from '../../../services/hotkeys.service';
 import {
   PixelGenerationDialog,
   GeneratePixelArtRequest,
 } from '../../../shared/pixel-generation-dialog/pixel-generation-dialog';
-import { PixelGenerationEngineService } from '../../../services/pixel-generation';
-import { EditorTransformService } from '../../../services/editor/editor-transform.service';
-import {
-  EditorFreeTransformService,
-  TransformHandle,
-} from '../../../services/editor/editor-free-transform.service';
-import {
-  EditorDistortService,
-  DistortHandle,
-} from '../../../services/editor/editor-distort.service';
-import {
-  EditorPerspectiveService,
-  PerspectiveHandle,
-} from '../../../services/editor/editor-perspective.service';
+import { EditorFreeTransformService } from '../../../services/editor/editor-free-transform.service';
+import { EditorDistortService } from '../../../services/editor/editor-distort.service';
+import { EditorPerspectiveService } from '../../../services/editor/editor-perspective.service';
 import { EditorWarpService } from '../../../services/editor/editor-warp.service';
 import { EditorPuppetWarpService } from '../../../services/editor/editor-puppet-warp.service';
-import {
-  FillSelectionDialog,
-  FillSelectionDialogResult,
-} from '../../../shared/components/fill-selection-dialog/fill-selection-dialog.component';
+import { FillSelectionDialog } from '../../../shared/components/fill-selection-dialog/fill-selection-dialog.component';
 import { FillSelectionService } from '../../../services/editor/fill-selection.service';
 import { ColorPickerStateService } from '../../../services/color-picker-state.service';
 import {
   CanvasViewportService,
   CanvasShapeService,
-  CanvasSelectionMoveService,
-  CanvasTransformStateService,
-  CanvasRenderService,
   CanvasGenerationService,
   ShapeDrawOptions,
+  CanvasFreeTransformHandler,
+  CanvasDistortHandler,
+  CanvasPerspectiveHandler,
+  CanvasWarpHandler,
+  CanvasPuppetWarpHandler,
+  CanvasContextMenuService,
+  ContextMenuActionId,
+  ContextMenuAction,
+  CanvasDrawService,
+  CanvasPointerService,
+  PointerState,
 } from '../../../services/editor/canvas';
-
-type ContextMenuActionId =
-  | 'deselect'
-  | 'invertSelection'
-  | 'growSelection'
-  | 'growBy1px'
-  | 'growBy2px'
-  | 'growBy5px'
-  | 'growCustom'
-  | 'makeCopyLayer'
-  | 'mergeVisibleToNewLayer'
-  | 'generate'
-  | 'generateFromLayer'
-  | 'generateFromVisible';
-
-interface ContextMenuAction {
-  id: ContextMenuActionId;
-  labelKey: string;
-  icon: string;
-  disabled?: boolean;
-  submenu?: ContextMenuAction[];
-}
 
 @Component({
   selector: 'pa-editor-canvas',
   templateUrl: './editor-canvas.component.html',
   styleUrls: ['./editor-canvas.component.css'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
   imports: [
     CommonModule,
     TranslocoPipe,
@@ -104,24 +67,25 @@ interface ContextMenuAction {
 export class EditorCanvas implements OnDestroy {
   private static readonly DEFAULT_WORKSPACE_WIDTH = 800;
   private static readonly DEFAULT_WORKSPACE_HEIGHT = 600;
+
   @ViewChild('canvasEl', { static: true })
   canvasEl!: ElementRef<HTMLCanvasElement>;
   @ViewChild('canvasContainer', { static: true })
   canvasContainer!: ElementRef<HTMLDivElement>;
-  private resizeObserver: ResizeObserver | null = null;
-  private workspaceWidth = signal(0);
-  private workspaceHeight = signal(0);
   @ViewChild(PixelGenerationDialog, { static: false })
   pixelGenerationDialog!: PixelGenerationDialog;
   @ViewChild(FillSelectionDialog, { static: false })
   fillSelectionDialog!: FillSelectionDialog;
+
+  private resizeObserver: ResizeObserver | null = null;
+  private workspaceWidth = signal(0);
+  private workspaceHeight = signal(0);
+
   readonly document = inject(EditorDocumentService);
   readonly documentSvc: EditorDocumentService = this.document;
   readonly tools = inject(EditorToolsService);
   readonly boneService = inject(EditorBoneService);
   readonly hotkeys = inject(HotkeysService);
-  readonly pixelGenEngine = inject(PixelGenerationEngineService);
-  readonly transformService = inject(EditorTransformService);
   readonly freeTransform = inject(EditorFreeTransformService);
   readonly distort = inject(EditorDistortService);
   readonly perspective = inject(EditorPerspectiveService);
@@ -131,10 +95,15 @@ export class EditorCanvas implements OnDestroy {
   readonly colorPickerState = inject(ColorPickerStateService);
   readonly viewport = inject(CanvasViewportService);
   readonly shapeService = inject(CanvasShapeService);
-  readonly selectionMoveService = inject(CanvasSelectionMoveService);
-  readonly transformState = inject(CanvasTransformStateService);
-  readonly renderService = inject(CanvasRenderService);
   readonly generationService = inject(CanvasGenerationService);
+  readonly freeTransformHandler = inject(CanvasFreeTransformHandler);
+  readonly distortHandler = inject(CanvasDistortHandler);
+  readonly perspectiveHandler = inject(CanvasPerspectiveHandler);
+  readonly warpHandler = inject(CanvasWarpHandler);
+  readonly puppetWarpHandler = inject(CanvasPuppetWarpHandler);
+  readonly contextMenu = inject(CanvasContextMenuService);
+  readonly drawService = inject(CanvasDrawService);
+  readonly pointerService = inject(CanvasPointerService);
 
   readonly mouseX = signal<number | null>(null);
   readonly mouseY = signal<number | null>(null);
@@ -147,141 +116,58 @@ export class EditorCanvas implements OnDestroy {
   get rotation() { return this.viewport.rotation; }
   get minScale() { return this.viewport.minScale; }
   get maxScale() { return this.viewport.maxScale; }
+
+  get contextMenuVisible() { return this.contextMenu.visible; }
+  get contextMenuPosition() { return this.contextMenu.position; }
+  get contextMenuActions() { return this.contextMenu.actions; }
+  get submenuVisible() { return this.contextMenu.submenuVisible; }
+  get submenuPosition() { return this.contextMenu.submenuPosition; }
+  get submenuActions() { return this.contextMenu.submenuActions; }
+  get inputDialogVisible() { return this.contextMenu.inputDialogVisible; }
+  get inputDialogPosition() { return this.contextMenu.inputDialogPosition; }
+  get inputDialogValue() { return this.contextMenu.inputDialogValue; }
+  get inputDialogTitle() { return this.contextMenu.inputDialogTitle; }
+  get inputDialogCallback() { return this.contextMenu.inputDialogCallback; }
+
   private readonly injector = inject(EnvironmentInjector);
   private readonly viewReady = signal(false);
   private readonly shapeStart = signal<{ x: number; y: number } | null>(null);
   private readonly shapeCurrent = signal<{ x: number; y: number } | null>(null);
-  private readonly activeShapeTool = signal<
-    'line' | 'circle' | 'square' | null
-  >(null);
+  private readonly activeShapeTool = signal<'line' | 'circle' | 'square' | null>(null);
   private readonly shapeConstrainUniform = signal(false);
   private readonly penPoints = signal<{ x: number; y: number }[]>([]);
   private readonly penDrawing = signal(false);
 
-  private panning = false;
-  private painting = false;
-  private lastPaintPos: { x: number; y: number } | null = null;
-  private selectionStart: { x: number; y: number } | null = null;
-  private selectionDragging = false;
-  private selectionMoving = false;
-  private selectionMoveStart: { x: number; y: number } | null = null;
-  private moveSelectionHotkeyActive = false;
-  private moveSelectionHotkeyParts = new Set<string>();
-  private selectionContentMoving = false;
-  private selectionContentMoveStart: { x: number; y: number } | null = null;
+  private pointerState: PointerState = {
+    panning: false,
+    painting: false,
+    lastPaintPos: null,
+    selectionStart: null,
+    selectionDragging: false,
+    selectionMoving: false,
+    selectionMoveStart: null,
+    selectionContentMoving: false,
+    selectionContentMoveStart: null,
+    shiftPressed: false,
+    lastPointer: { x: 0, y: 0 },
+    shaping: false,
+    currentBoneId: null,
+    draggingPointId: null,
+    draggingPointBoneId: null,
+  };
+
   private movingContentBuffer: string[] | null = null;
-  private movingContentOriginalRect: {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-  } | null = null;
+  private movingContentOriginalRect: { x: number; y: number; width: number; height: number } | null = null;
   private originalLayerId: string | null = null;
-  private freeTransformOriginalBuffer: string[] | null = null;
-  private freeTransformOriginalRect: {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-  } | null = null;
-  private freeTransformLayerId: string | null = null;
-  private freeTransformPreviewIndices: number[] | null = null;
-  private freeTransformPreviewBackup: Map<number, string> | null = null;
-  private freeTransformDuplicate = false;
-  private freeTransformMirrorX = false;
-  private freeTransformMirrorY = false;
-  private distortOriginalBuffer: string[] | null = null;
-  private distortOriginalRect: {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-  } | null = null;
-  private distortLayerId: string | null = null;
-  private distortFullLayerBackup: string[] | null = null;
-  private perspectiveOriginalBuffer: string[] | null = null;
-  private perspectiveOriginalRect: {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-  } | null = null;
-  private perspectiveLayerId: string | null = null;
-  private perspectiveFullLayerBackup: string[] | null = null;
-  private warpOriginalBuffer: string[] | null = null;
-  private warpOriginalRect: {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-  } | null = null;
-  private warpLayerId: string | null = null;
-  private warpFullLayerBackup: string[] | null = null;
-  private puppetWarpOriginalBuffer: string[] | null = null;
-  private puppetWarpOriginalRect: {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-  } | null = null;
-  private puppetWarpLayerId: string | null = null;
-  private puppetWarpFullLayerBackup: string[] | null = null;
-  private shiftPressed = false;
-  private lastPointer = { x: 0, y: 0 };
-  private shaping = false;
   private stopRenderEffect: EffectRef | null = null;
-  private currentBoneId: string | null = null;
-  private draggingPointId: string | null = null;
-  private draggingPointBoneId: string | null = null;
   readonly tileSize = signal(1);
-  readonly contextMenuVisible = signal(false);
-  readonly contextMenuPosition = signal<{ x: number; y: number }>({
-    x: 0,
-    y: 0,
-  });
-  readonly contextMenuActions = signal<ContextMenuAction[]>([]);
-  readonly submenuVisible = signal(false);
-  readonly submenuPosition = signal<{ x: number; y: number }>({ x: 0, y: 0 });
-  readonly submenuActions = signal<ContextMenuAction[]>([]);
-  readonly inputDialogVisible = signal(false);
-  readonly inputDialogPosition = signal<{ x: number; y: number }>({
-    x: 0,
-    y: 0,
-  });
-  readonly inputDialogValue = signal('10');
-  readonly inputDialogTitle = signal('');
-  readonly inputDialogCallback = signal<((value: string) => void) | null>(null);
   private resizeListener: (() => void) | null = null;
   private keyListener: ((e: KeyboardEvent) => void) | null = null;
   private readonly defaultCursor = `url('/cursors/link.png') 12 12, link`;
   private readonly brushCursor = `url('/cursors/handwriting.png') 12 12, crosshair`;
   private readonly eraserCursor = `url('/cursors/unavailable.png') 12 12, cell`;
   private readonly handGrabbingCursor = `url('/cursors/grab.png') 12 12, grab`;
-  private readonly bayer4 = [
-    [0, 8, 2, 10],
-    [12, 4, 14, 6],
-    [3, 11, 1, 9],
-    [15, 7, 13, 5],
-  ];
-  private readonly gradientSteps = 8;
-
-  private capturePointer(ev: PointerEvent): void {
-    const target = ev.currentTarget as HTMLElement;
-    if (target?.setPointerCapture) {
-      target.setPointerCapture(ev.pointerId);
-    }
-  }
-
-  private releasePointer(ev: PointerEvent): void {
-    const target = ev.currentTarget as HTMLElement;
-    if (target?.releasePointerCapture) {
-      try {
-        if (target.hasPointerCapture(ev.pointerId)) {
-          target.releasePointerCapture(ev.pointerId);
-        }
-      } catch {}
-    }
-  }
+  infoVisible = signal(true);
 
   constructor() {
     this.stopRenderEffect = effect(() => {
@@ -289,33 +175,32 @@ export class EditorCanvas implements OnDestroy {
       return null as any;
     });
 
-    // If selection is cleared while free transform is active, cancel and clean state.
     effect(() => {
       const sel = this.document.selectionRect();
       const active = this.freeTransform.isActive();
       if (active && (!sel || sel.width <= 0 || sel.height <= 0)) {
-        this.cancelFreeTransform();
+        this.freeTransformHandler.cancel();
       }
     });
 
     effect(() => {
       const tool = this.tools.currentTool();
-      if (tool !== 'bone' && this.currentBoneId) {
-        this.currentBoneId = null;
+      if (tool !== 'bone' && this.pointerState.currentBoneId) {
+        this.pointerState.currentBoneId = null;
       }
     });
 
     effect(() => {
       const active = this.freeTransform.isActive();
-      if (active && !this.freeTransformOriginalBuffer) {
-        this.activateFreeTransform();
+      if (active && !this.freeTransformHandler.hasOriginalBuffer()) {
+        this.freeTransformHandler.activate();
       }
     });
 
     effect(() => {
       const active = this.distort.isActive();
-      if (active && !this.distortOriginalBuffer) {
-        this.activateDistort();
+      if (active && !this.distortHandler.hasOriginalBuffer()) {
+        this.distortHandler.activate();
       }
     });
 
@@ -323,14 +208,14 @@ export class EditorCanvas implements OnDestroy {
       const sel = this.document.selectionRect();
       const active = this.distort.isActive();
       if (active && (!sel || sel.width <= 0 || sel.height <= 0)) {
-        this.cancelDistort();
+        this.distortHandler.cancel();
       }
     });
 
     effect(() => {
       const active = this.perspective.isActive();
-      if (active && !this.perspectiveOriginalBuffer) {
-        this.activatePerspective();
+      if (active && !this.perspectiveHandler.hasOriginalBuffer()) {
+        this.perspectiveHandler.activate();
       }
     });
 
@@ -338,14 +223,14 @@ export class EditorCanvas implements OnDestroy {
       const sel = this.document.selectionRect();
       const active = this.perspective.isActive();
       if (active && (!sel || sel.width <= 0 || sel.height <= 0)) {
-        this.cancelPerspective();
+        this.perspectiveHandler.cancel();
       }
     });
 
     effect(() => {
       const active = this.warp.isActive();
-      if (active && !this.warpOriginalBuffer) {
-        this.activateWarp();
+      if (active && !this.warpHandler.hasOriginalBuffer()) {
+        this.warpHandler.activate();
       }
     });
 
@@ -353,14 +238,14 @@ export class EditorCanvas implements OnDestroy {
       const sel = this.document.selectionRect();
       const active = this.warp.isActive();
       if (active && (!sel || sel.width <= 0 || sel.height <= 0)) {
-        this.cancelWarp();
+        this.warpHandler.cancel();
       }
     });
 
     effect(() => {
       const active = this.puppetWarp.isActive();
-      if (active && !this.puppetWarpOriginalBuffer) {
-        this.activatePuppetWarp();
+      if (active && !this.puppetWarpHandler.hasOriginalBuffer()) {
+        this.puppetWarpHandler.activate();
       }
     });
 
@@ -368,7 +253,7 @@ export class EditorCanvas implements OnDestroy {
       const sel = this.document.selectionRect();
       const active = this.puppetWarp.isActive();
       if (active && (!sel || sel.width <= 0 || sel.height <= 0)) {
-        this.cancelPuppetWarp();
+        this.puppetWarpHandler.cancel();
       }
     });
 
@@ -438,9 +323,7 @@ export class EditorCanvas implements OnDestroy {
       defaultKey: 'shift+t',
       handler: () => {
         const sel = this.document.selectionRect();
-        if (!sel || sel.width <= 0 || sel.height <= 0) {
-          return;
-        }
+        if (!sel || sel.width <= 0 || sel.height <= 0) return;
         this.freeTransform.startTransform(sel.x, sel.y, sel.width, sel.height);
       },
     });
@@ -455,8 +338,7 @@ export class EditorCanvas implements OnDestroy {
 
   async openFillSelectionDialog() {
     const selectionRect = this.document.selectionRect();
-    if (!selectionRect || selectionRect.width <= 0 || selectionRect.height <= 0)
-      return;
+    if (!selectionRect || selectionRect.width <= 0 || selectionRect.height <= 0) return;
 
     const result = await this.fillSelectionDialog?.open();
     if (!result) return;
@@ -473,63 +355,19 @@ export class EditorCanvas implements OnDestroy {
 
     if (success) {
       this.tools.setFillMode(result.mode);
-      if (result.color) {
-        this.tools.setFillColor(result.color);
-      }
-      if (result.patternId) {
-        this.tools.setFillPatternId(result.patternId);
-      }
-      if (result.gradientStartColor) {
-        this.tools.setFillGradientStartColor(result.gradientStartColor);
-      }
-      if (result.gradientEndColor) {
-        this.tools.setFillGradientEndColor(result.gradientEndColor);
-      }
-      if (result.gradientType) {
-        this.tools.setFillGradientType(result.gradientType);
-      }
-      if (result.gradientAngle !== undefined) {
-        this.tools.setFillGradientAngle(result.gradientAngle);
-      }
+      if (result.color) this.tools.setFillColor(result.color);
+      if (result.patternId) this.tools.setFillPatternId(result.patternId);
+      if (result.gradientStartColor) this.tools.setFillGradientStartColor(result.gradientStartColor);
+      if (result.gradientEndColor) this.tools.setFillGradientEndColor(result.gradientEndColor);
+      if (result.gradientType) this.tools.setFillGradientType(result.gradientType);
+      if (result.gradientAngle !== undefined) this.tools.setFillGradientAngle(result.gradientAngle);
     }
   }
 
   private activateMoveSelectionHotkey() {
     const binding = this.hotkeys.getBinding('edit.moveSelectionPixels');
     if (!binding) return;
-    this.moveSelectionHotkeyActive = true;
-    this.moveSelectionHotkeyParts.clear();
-    for (const part of binding.split('+')) {
-      if (part) {
-        this.moveSelectionHotkeyParts.add(part);
-      }
-    }
-  }
-
-  private deactivateMoveSelectionHotkey() {
-    this.moveSelectionHotkeyActive = false;
-    this.moveSelectionHotkeyParts.clear();
-  }
-
-  private handleMoveSelectionHotkeyRelease(event: KeyboardEvent) {
-    if (!this.moveSelectionHotkeyActive) {
-      return;
-    }
-    const part = this.normalizeMoveSelectionKey(event.key);
-    if (part && this.moveSelectionHotkeyParts.has(part)) {
-      this.deactivateMoveSelectionHotkey();
-    }
-  }
-
-  private normalizeMoveSelectionKey(key: string): string | null {
-    const lower = key.toLowerCase();
-    if (lower === 'control' || lower === 'meta') {
-      return 'ctrl';
-    }
-    if (lower === 'shift' || lower === 'alt') {
-      return lower;
-    }
-    return lower;
+    this.pointerService.activateMoveSelectionHotkey(binding);
   }
 
   private readonly layoutEffect = effect(
@@ -538,29 +376,21 @@ export class EditorCanvas implements OnDestroy {
       const w = this.document.canvasWidth();
       const h = this.document.canvasHeight();
       this.viewport.updateMaxScale(w, h);
-      const scheduler =
-        typeof queueMicrotask === 'function'
-          ? queueMicrotask
-          : (cb: () => void) => Promise.resolve().then(cb);
+      const scheduler = typeof queueMicrotask === 'function'
+        ? queueMicrotask
+        : (cb: () => void) => Promise.resolve().then(cb);
       scheduler(() => this.centerAndFitCanvas());
     },
     { injector: this.injector },
   );
 
-  // compute the CSS cursor for the canvas based on current tool and state
   cursor(): string {
-    if (this.panning) return this.handGrabbingCursor;
+    if (this.pointerState.panning) return this.handGrabbingCursor;
     const tool = this.tools.currentTool();
-    if (
-      tool === 'rect-select' ||
-      tool === 'ellipse-select' ||
-      tool === 'lasso-select'
-    )
-      return `crosshair`;
+    if (tool === 'rect-select' || tool === 'ellipse-select' || tool === 'lasso-select') return `crosshair`;
     if (tool === 'brush') return this.brushCursor;
     if (tool === 'eraser') return this.eraserCursor;
-    if (tool === 'line' || tool === 'circle' || tool === 'square' || tool === 'pen')
-      return `crosshair`;
+    if (tool === 'line' || tool === 'circle' || tool === 'square' || tool === 'pen') return `crosshair`;
     if (tool === 'bone') return `crosshair`;
     return this.defaultCursor;
   }
@@ -584,145 +414,25 @@ export class EditorCanvas implements OnDestroy {
     }
 
     this.centerAndFitCanvas();
-    if (
-      typeof window !== 'undefined' &&
-      typeof window.requestAnimationFrame === 'function'
-    ) {
+    if (typeof window !== 'undefined' && typeof window.requestAnimationFrame === 'function') {
       window.requestAnimationFrame(() => this.centerAndFitCanvas());
     }
     this.updateTileSize(this.tools.brushSize());
 
     const flatLayers = this.document.getFlattenedLayers();
     for (const l of flatLayers) {
-      this.document.ensureLayerBuffer(
-        l.id,
-        this.document.canvasWidth(),
-        this.document.canvasHeight(),
-      );
+      this.document.ensureLayerBuffer(l.id, this.document.canvasWidth(), this.document.canvasHeight());
     }
 
     this.resizeListener = () => this.centerAndFitCanvas();
     if (typeof window !== 'undefined') {
       window.addEventListener('resize', this.resizeListener as EventListener);
-      // listen for Escape to cancel in-progress lasso selections
-      this.keyListener = (ev: KeyboardEvent) => {
-        const target = ev.target as HTMLElement | null;
-        if (target) {
-          const tag = target.tagName.toLowerCase();
-          if (
-            tag === 'input' ||
-            tag === 'textarea' ||
-            target.isContentEditable
-          ) {
-            return;
-          }
-        }
-        if (ev.key === 'Escape') {
-          if (this.freeTransform.isActive()) {
-            this.cancelFreeTransform();
-            return;
-          }
-          if (this.distort.isActive()) {
-            this.cancelDistort();
-            return;
-          }
-          if (this.perspective.isActive()) {
-            this.cancelPerspective();
-            return;
-          }
-          if (this.contextMenuVisible()) {
-            this.closeContextMenu();
-            return;
-          }
-          const tool = this.tools.currentTool();
-          if (tool === 'pen' && this.penDrawing()) {
-            this.cancelPenPath();
-            return;
-          }
-          if (tool === 'bone') {
-            this.currentBoneId = null;
-            this.boneService.clearSelection();
-            return;
-          }
-          if (tool === 'lasso-select') {
-            this.document.selectionPolygon.set(null as any);
-            this.document.selectionRect.set(null as any);
-            this.document.selectionShape.set('rect');
-            this.selectionDragging = false;
-            this.selectionStart = null;
-          }
-          return;
-        }
-        if (ev.key === 'Enter') {
-          if (this.freeTransform.isActive()) {
-            this.commitFreeTransform();
-            return;
-          }
-          if (this.distort.isActive()) {
-            this.commitDistort();
-            return;
-          }
-          if (this.perspective.isActive()) {
-            this.commitPerspective();
-            return;
-          }
-          const tool = this.tools.currentTool();
-          if (tool === 'pen' && this.penDrawing()) {
-            this.finishPenPath();
-            return;
-          }
-        }
-        const key = ev.key?.toLowerCase?.() ?? ev.key;
-        if (ev.ctrlKey && ev.shiftKey && key === 'd') {
-          ev.preventDefault();
-          this.document.clearSelection();
-          return;
-        }
-        if (ev.ctrlKey && key === 'd') {
-          ev.preventDefault();
-          this.document.duplicateLayer();
-          return;
-        }
-        if (ev.ctrlKey && key === 'l') {
-          ev.preventDefault();
-          const selectedId = this.document.selectedLayerId();
-          if (selectedId) {
-            this.document.toggleLayerLock(selectedId);
-          }
-          return;
-        }
-        if (ev.key === 'Delete') {
-          ev.preventDefault();
-          const selectedIds = Array.from(this.document.selectedLayerIds());
-          for (const id of selectedIds) {
-            this.document.removeLayer(id);
-          }
-          return;
-        }
-        if (ev.ctrlKey && ev.shiftKey && key === 'g') {
-          ev.preventDefault();
-          const selectedIds = Array.from(this.document.selectedLayerIds());
-          if (selectedIds.length >= 2) {
-            this.document.ungroupLayers(selectedIds[0]);
-          }
-          return;
-        }
-        // Ctrl+T is browser-reserved (new tab) and cannot be reliably overridden.
-        // Free Transform is bound to Shift+T via HotkeysService.
-        if (ev.ctrlKey && key === 'g') {
-          ev.preventDefault();
-          const selectedIds = Array.from(this.document.selectedLayerIds());
-          if (selectedIds.length >= 2) {
-            this.document.groupLayers(selectedIds);
-          }
-          return;
-        }
-      };
+      this.keyListener = (ev: KeyboardEvent) => this.handleKeyDown(ev);
       window.addEventListener('keydown', this.keyListener as EventListener);
       window.addEventListener('keyup', (ev: KeyboardEvent) => {
-        this.handleMoveSelectionHotkeyRelease(ev);
+        this.pointerService.handleMoveSelectionHotkeyRelease(ev);
         if (ev.key === 'Shift') {
-          this.shiftPressed = false;
+          this.pointerState.shiftPressed = false;
           this.freeTransform.setSnapRotation(false);
           this.freeTransform.setConstrainProportions(false);
         }
@@ -732,142 +442,97 @@ export class EditorCanvas implements OnDestroy {
     this.viewReady.set(true);
   }
 
+  private handleKeyDown(ev: KeyboardEvent): void {
+    const target = ev.target as HTMLElement | null;
+    if (target) {
+      const tag = target.tagName.toLowerCase();
+      if (tag === 'input' || tag === 'textarea' || target.isContentEditable) return;
+    }
+
+    if (ev.key === 'Escape') {
+      if (this.freeTransform.isActive()) { this.freeTransformHandler.cancel(); return; }
+      if (this.distort.isActive()) { this.distortHandler.cancel(); return; }
+      if (this.perspective.isActive()) { this.perspectiveHandler.cancel(); return; }
+      if (this.contextMenuVisible()) { this.closeContextMenu(); return; }
+      const tool = this.tools.currentTool();
+      if (tool === 'pen' && this.penDrawing()) { this.cancelPenPath(); return; }
+      if (tool === 'bone') {
+        this.pointerState.currentBoneId = null;
+        this.boneService.clearSelection();
+        return;
+      }
+      if (tool === 'lasso-select') {
+        this.document.selectionPolygon.set(null as any);
+        this.document.selectionRect.set(null as any);
+        this.document.selectionShape.set('rect');
+        this.pointerState.selectionDragging = false;
+        this.pointerState.selectionStart = null;
+      }
+      return;
+    }
+
+    if (ev.key === 'Enter') {
+      if (this.freeTransform.isActive()) { this.freeTransformHandler.commit(); return; }
+      if (this.distort.isActive()) { this.distortHandler.commit(); return; }
+      if (this.perspective.isActive()) { this.perspectiveHandler.commit(); return; }
+      const tool = this.tools.currentTool();
+      if (tool === 'pen' && this.penDrawing()) { this.finishPenPath(); return; }
+    }
+
+    const key = ev.key?.toLowerCase?.() ?? ev.key;
+    if (ev.ctrlKey && ev.shiftKey && key === 'd') { ev.preventDefault(); this.document.clearSelection(); return; }
+    if (ev.ctrlKey && key === 'd') { ev.preventDefault(); this.document.duplicateLayer(); return; }
+    if (ev.ctrlKey && key === 'l') {
+      ev.preventDefault();
+      const selectedId = this.document.selectedLayerId();
+      if (selectedId) this.document.toggleLayerLock(selectedId);
+      return;
+    }
+    if (ev.key === 'Delete') {
+      ev.preventDefault();
+      const selectedIds = Array.from(this.document.selectedLayerIds());
+      for (const id of selectedIds) this.document.removeLayer(id);
+      return;
+    }
+    if (ev.ctrlKey && ev.shiftKey && key === 'g') {
+      ev.preventDefault();
+      const selectedIds = Array.from(this.document.selectedLayerIds());
+      if (selectedIds.length >= 2) this.document.ungroupLayers(selectedIds[0]);
+      return;
+    }
+    if (ev.ctrlKey && key === 'g') {
+      ev.preventDefault();
+      const selectedIds = Array.from(this.document.selectedLayerIds());
+      if (selectedIds.length >= 2) this.document.groupLayers(selectedIds);
+      return;
+    }
+  }
+
   updateTileSize(brushSize = 1, desiredScreenTilePx = 24) {
     const s = Math.max(0.001, this.scale());
-    const tile = Math.max(
-      1,
-      Math.round(desiredScreenTilePx / (s * Math.max(1, brushSize))),
-    );
+    const tile = Math.max(1, Math.round(desiredScreenTilePx / (s * Math.max(1, brushSize))));
     this.tileSize.set(tile);
   }
 
   onPointerMove(ev: PointerEvent) {
     const rect = this.canvasEl.nativeElement.getBoundingClientRect();
+    const callbacks = this.createPointerCallbacks();
+    const result = this.pointerService.handlePointerMove(ev, rect, this.pointerState, callbacks);
+    this.mouseX.set(result.mouseX);
+    this.mouseY.set(result.mouseY);
+    this.hoverX.set(result.hoverX);
+    this.hoverY.set(result.hoverY);
 
-    const visX = ev.clientX - rect.left;
-    const visY = ev.clientY - rect.top;
-    this.mouseX.set(Math.round(visX));
-    this.mouseY.set(Math.round(visY));
-
-    const w = this.document.canvasWidth();
-    const h = this.document.canvasHeight();
-    const currentScale = this.scale();
-    const currentPanX = this.panX();
-    const currentPanY = this.panY();
-    const logicalX = Math.floor((visX - currentPanX) / currentScale);
-    const logicalY = Math.floor((visY - currentPanY) / currentScale);
-
-    if (logicalX >= 0 && logicalX < w && logicalY >= 0 && logicalY < h) {
-      this.hoverX.set(logicalX);
-      this.hoverY.set(logicalY);
-    } else {
-      this.hoverX.set(null);
-      this.hoverY.set(null);
-    }
-
-    if (this.panning) {
-      const dx = ev.clientX - this.lastPointer.x;
-      const dy = ev.clientY - this.lastPointer.y;
-      this.panX.set(this.panX() + dx);
-      this.panY.set(this.panY() + dy);
-      this.lastPointer.x = ev.clientX;
-      this.lastPointer.y = ev.clientY;
-    }
-
-    if (this.freeTransform.isDraggingHandle()) {
-      const handle = this.freeTransform.isDraggingHandle();
-      if (handle === 'rotate-center') {
-        this.freeTransform.setSnapRotation(ev.shiftKey || this.shiftPressed);
-        this.freeTransform.setConstrainProportions(false);
-      } else {
-        this.freeTransform.setConstrainProportions(
-          ev.shiftKey || this.shiftPressed,
-        );
-        this.freeTransform.setSnapRotation(false);
-      }
-      this.freeTransform.updateHandleDrag(logicalX, logicalY);
-      this.renderLiveFreeTransformPreview();
-      return;
-    }
-
-    if (this.distort.isDraggingHandle()) {
-      this.distort.updateHandleDrag(logicalX, logicalY);
-      this.renderLiveDistortPreview();
-      return;
-    }
-
-    if (this.perspective.isDraggingHandle()) {
-      this.perspective.updateHandleDrag(logicalX, logicalY);
-      this.renderLivePerspectivePreview();
-      return;
-    }
-
-    if (this.warp.isDraggingNode()) {
-      this.warp.updateNodeDrag(logicalX, logicalY);
-      this.renderLiveWarpPreview();
-      return;
-    }
-
-    if (this.puppetWarp.isDraggingPin()) {
-      this.puppetWarp.updatePinDrag(logicalX, logicalY);
-      this.renderLivePuppetWarpPreview();
-      return;
-    }
-
-    if (this.selectionMoving && this.selectionMoveStart) {
-      if (logicalX < 0 || logicalX >= w || logicalY < 0 || logicalY >= h) {
-        return;
-      }
-      const dx = logicalX - this.selectionMoveStart.x;
-      const dy = logicalY - this.selectionMoveStart.y;
-      if (dx !== 0 || dy !== 0) {
-        this.document.moveSelection(dx, dy);
-        this.selectionMoveStart = { x: logicalX, y: logicalY };
-      }
-      return;
-    }
-
-    if (this.selectionContentMoving && this.selectionContentMoveStart) {
-      if (logicalX < 0 || logicalX >= w || logicalY < 0 || logicalY >= h) {
-        return;
-      }
-      const dx = logicalX - this.selectionContentMoveStart.x;
-      const dy = logicalY - this.selectionContentMoveStart.y;
-      if (dx !== 0 || dy !== 0) {
-        this.moveSelectionContent(dx, dy);
-        this.selectionContentMoveStart = { x: logicalX, y: logicalY };
-      }
-      return;
-    }
-
-    if (this.selectionDragging) {
-      const clampedX = Math.max(0, Math.min(w - 1, logicalX));
-      const clampedY = Math.max(0, Math.min(h - 1, logicalY));
-      const tool = this.tools.currentTool();
-      if (tool === 'lasso-select') {
-        this.document.addLassoPoint(clampedX, clampedY);
-      } else {
-        let endX = clampedX;
-        let endY = clampedY;
-        if (ev.shiftKey && this.selectionStart) {
-          const dx = clampedX - this.selectionStart.x;
-          const dy = clampedY - this.selectionStart.y;
-          const absDx = Math.abs(dx);
-          const absDy = Math.abs(dy);
-          const max = Math.max(absDx, absDy);
-          const sx = dx >= 0 ? 1 : -1;
-          const sy = dy >= 0 ? 1 : -1;
-          endX = this.selectionStart.x + sx * max;
-          endY = this.selectionStart.y + sy * max;
-          endX = Math.max(0, Math.min(endX, w - 1));
-          endY = Math.max(0, Math.min(endY, h - 1));
-        }
-        this.document.updateSelection(endX, endY);
-      }
-      return;
-    }
-
-    if (this.shaping) {
+    if (this.pointerState.shaping) {
+      const w = this.document.canvasWidth();
+      const h = this.document.canvasHeight();
+      const currentScale = this.scale();
+      const currentPanX = this.panX();
+      const currentPanY = this.panY();
+      const visX = ev.clientX - rect.left;
+      const visY = ev.clientY - rect.top;
+      const logicalX = Math.floor((visX - currentPanX) / currentScale);
+      const logicalY = Math.floor((visY - currentPanY) / currentScale);
       const clampedX = Math.max(0, Math.min(w - 1, logicalX));
       const clampedY = Math.max(0, Math.min(h - 1, logicalY));
       const active = this.activeShapeTool();
@@ -877,62 +542,6 @@ export class EditorCanvas implements OnDestroy {
         this.shapeConstrainUniform.set(false);
       }
       this.shapeCurrent.set({ x: clampedX, y: clampedY });
-    }
-
-    if (this.draggingPointId && this.draggingPointBoneId) {
-      const clampedX = Math.max(0, Math.min(w - 1, logicalX));
-      const clampedY = Math.max(0, Math.min(h - 1, logicalY));
-      const frameId = this.getCurrentFrameId();
-      const currentAnim = this.document.getCurrentAnimation();
-      const animationId = currentAnim?.id;
-      const currentTime = this.document.keyframeService?.getCurrentTime();
-      this.boneService.updatePoint(
-        frameId,
-        this.draggingPointBoneId,
-        this.draggingPointId,
-        clampedX,
-        clampedY,
-        animationId,
-        currentTime,
-      );
-      return;
-    }
-
-    if (this.painting) {
-      const selectedLayer = this.document.selectedLayer();
-      if (selectedLayer?.locked) {
-        return;
-      }
-      const clampedX = Math.max(0, Math.min(w - 1, logicalX));
-      const clampedY = Math.max(0, Math.min(h - 1, logicalY));
-      const layerId = this.document.selectedLayerId();
-      const tool = this.tools.currentTool();
-      const color = tool === 'eraser' ? null : this.tools.brushColor();
-      const size =
-        tool === 'eraser' ? this.tools.eraserSize() : this.tools.brushSize();
-      if (this.lastPaintPos) {
-        this.drawLinePaint(
-          layerId,
-          this.lastPaintPos.x,
-          this.lastPaintPos.y,
-          clampedX,
-          clampedY,
-          size,
-          color,
-        );
-      } else {
-        this.document.applyBrushToLayer(
-          layerId,
-          clampedX,
-          clampedY,
-          size,
-          color,
-          tool === 'eraser'
-            ? { eraserStrength: this.tools.eraserStrength() }
-            : undefined,
-        );
-      }
-      this.lastPaintPos = { x: clampedX, y: clampedY };
     }
   }
 
@@ -948,29 +557,18 @@ export class EditorCanvas implements OnDestroy {
     const target = ev.target as HTMLElement | null;
     if (target) {
       const tag = target.tagName.toLowerCase();
-      if (tag === 'input' || tag === 'textarea' || target.isContentEditable)
-        return;
+      if (tag === 'input' || tag === 'textarea' || target.isContentEditable) return;
     }
 
     const rect = container.getBoundingClientRect();
-    if (
-      ev.clientX < rect.left ||
-      ev.clientX > rect.right ||
-      ev.clientY < rect.top ||
-      ev.clientY > rect.bottom
-    ) {
-      return;
-    }
+    if (ev.clientX < rect.left || ev.clientX > rect.right || ev.clientY < rect.top || ev.clientY > rect.bottom) return;
 
     ev.preventDefault();
     ev.stopPropagation();
 
     let delta = ev.deltaY;
-    if (ev.deltaMode === WheelEvent.DOM_DELTA_LINE) {
-      delta *= 16;
-    } else if (ev.deltaMode === WheelEvent.DOM_DELTA_PAGE) {
-      delta *= rect.height;
-    }
+    if (ev.deltaMode === WheelEvent.DOM_DELTA_LINE) delta *= 16;
+    else if (ev.deltaMode === WheelEvent.DOM_DELTA_PAGE) delta *= rect.height;
 
     const zoomIntensity = 0.002;
     const factor = Math.exp(-delta * zoomIntensity);
@@ -979,569 +577,49 @@ export class EditorCanvas implements OnDestroy {
   }
 
   onPointerDown(ev: PointerEvent) {
-    if (this.contextMenuVisible()) {
-      this.closeContextMenu();
-    }
-    if (ev.button === 1 || ev.ctrlKey) {
-      this.panning = true;
-      this.lastPointer.x = ev.clientX;
-      this.lastPointer.y = ev.clientY;
-      this.capturePointer(ev);
-    }
-
+    if (this.contextMenuVisible()) this.closeContextMenu();
     const rect = this.canvasEl.nativeElement.getBoundingClientRect();
-    const visX = ev.clientX - rect.left;
-    const visY = ev.clientY - rect.top;
-    const w = this.document.canvasWidth();
-    const h = this.document.canvasHeight();
-    const currentScale = this.scale();
-    const currentPanX = this.panX();
-    const currentPanY = this.panY();
-    const logicalX = Math.floor((visX - currentPanX) / currentScale);
-    const logicalY = Math.floor((visY - currentPanY) / currentScale);
-    const tool = this.tools.currentTool();
-    const insideCanvas =
-      logicalX >= 0 && logicalX < w && logicalY >= 0 && logicalY < h;
+    const callbacks = this.createPointerCallbacks();
+    this.pointerService.handlePointerDown(
+      ev,
+      rect,
+      this.pointerState,
+      callbacks,
+      (val) => this.shapeConstrainUniform.set(val),
+    );
+  }
 
-    if (ev.button === 2) {
-      return;
-    }
+  onPointerUp(ev: PointerEvent) {
+    const callbacks = this.createPointerCallbacks();
+    this.pointerService.handlePointerUp(ev, this.pointerState, callbacks);
+  }
 
-    if (ev.button === 0) {
-      const distortState = this.distort.distortState();
-      if (distortState) {
-        const handleSize = Math.max(
-          3,
-          Math.round(5 / Math.max(0.001, this.scale())),
-        );
+  private createPointerCallbacks() {
+    return {
+      capturePointer: (ev: PointerEvent) => this.capturePointer(ev),
+      releasePointer: (ev: PointerEvent) => this.releasePointer(ev),
+      startShape: (mode: 'line' | 'circle' | 'square', x: number, y: number) => this.startShape(mode, x, y),
+      finishShape: (shiftKey: boolean) => this.finishShape(shiftKey),
+      addPenPoint: (x: number, y: number) => this.addPenPoint(x, y),
+      getCurrentFrameId: () => this.getCurrentFrameId(),
+      startSelectionContentMove: () => this.startSelectionContentMove(),
+      moveSelectionContent: (dx: number, dy: number) => this.moveSelectionContent(dx, dy),
+      endSelectionContentMove: () => this.endSelectionContentMove(),
+      isPointInSelection: (x: number, y: number) => this.isPointInSelection(x, y),
+      drawLinePaint: (layerId: string, x0: number, y0: number, x1: number, y1: number, brushSize: number, color: string | null) =>
+        this.drawLinePaint(layerId, x0, y0, x1, y1, brushSize, color),
+    };
+  }
 
-        const buttonPositions = this.computeDistortButtonPositions(
-          distortState,
-          this.scale(),
-        );
-        const btnSize = buttonPositions.btnSize;
+  private capturePointer(ev: PointerEvent): void {
+    const target = ev.currentTarget as HTMLElement;
+    if (target?.setPointerCapture) target.setPointerCapture(ev.pointerId);
+  }
 
-        if (
-          logicalX >= buttonPositions.commitX &&
-          logicalX <= buttonPositions.commitX + btnSize &&
-          logicalY >= buttonPositions.commitY &&
-          logicalY <= buttonPositions.commitY + btnSize
-        ) {
-          this.commitDistort();
-          return;
-        }
-
-        if (
-          logicalX >= buttonPositions.cancelX &&
-          logicalX <= buttonPositions.cancelX + btnSize &&
-          logicalY >= buttonPositions.cancelY &&
-          logicalY <= buttonPositions.cancelY + btnSize
-        ) {
-          this.cancelDistort();
-          return;
-        }
-
-        const handles: DistortHandle[] = [
-          'top-left',
-          'top-right',
-          'bottom-right',
-          'bottom-left',
-        ];
-
-        for (const handle of handles) {
-          const pos = this.distort.getHandlePosition(handle);
-          if (pos) {
-            const dx = logicalX - pos.x;
-            const dy = logicalY - pos.y;
-            if (dx * dx + dy * dy <= handleSize * handleSize) {
-              this.capturePointer(ev);
-              this.distort.startHandleDrag(handle, logicalX, logicalY);
-              return;
-            }
-          }
-        }
-
-        return;
-      }
-
-      const perspectiveState = this.perspective.perspectiveState();
-      if (perspectiveState) {
-        const handleSize = Math.max(
-          3,
-          Math.round(5 / Math.max(0.001, this.scale())),
-        );
-
-        const buttonPositions = this.computePerspectiveButtonPositions(
-          perspectiveState,
-          this.scale(),
-        );
-        const btnSize = buttonPositions.btnSize;
-
-        if (
-          logicalX >= buttonPositions.commitX &&
-          logicalX <= buttonPositions.commitX + btnSize &&
-          logicalY >= buttonPositions.commitY &&
-          logicalY <= buttonPositions.commitY + btnSize
-        ) {
-          this.commitPerspective();
-          return;
-        }
-
-        if (
-          logicalX >= buttonPositions.cancelX &&
-          logicalX <= buttonPositions.cancelX + btnSize &&
-          logicalY >= buttonPositions.cancelY &&
-          logicalY <= buttonPositions.cancelY + btnSize
-        ) {
-          this.cancelPerspective();
-          return;
-        }
-
-        const handles: PerspectiveHandle[] = [
-          'top-left',
-          'top-right',
-          'bottom-right',
-          'bottom-left',
-        ];
-
-        for (const handle of handles) {
-          const pos = this.perspective.getHandlePosition(handle);
-          if (pos) {
-            const dx = logicalX - pos.x;
-            const dy = logicalY - pos.y;
-            if (dx * dx + dy * dy <= handleSize * handleSize) {
-              this.capturePointer(ev);
-              this.perspective.startHandleDrag(handle, logicalX, logicalY);
-              return;
-            }
-          }
-        }
-
-        return;
-      }
-
-      const warpState = this.warp.warpState();
-      if (warpState) {
-        const handleSize = Math.max(
-          3,
-          Math.round(5 / Math.max(0.001, this.scale())),
-        );
-
-        const buttonPositions = this.computeWarpButtonPositions(
-          warpState,
-          this.scale(),
-        );
-        const btnSize = buttonPositions.btnSize;
-
-        if (
-          logicalX >= buttonPositions.commitX &&
-          logicalX <= buttonPositions.commitX + btnSize &&
-          logicalY >= buttonPositions.commitY &&
-          logicalY <= buttonPositions.commitY + btnSize
-        ) {
-          this.commitWarp();
-          return;
-        }
-
-        if (
-          logicalX >= buttonPositions.cancelX &&
-          logicalX <= buttonPositions.cancelX + btnSize &&
-          logicalY >= buttonPositions.cancelY &&
-          logicalY <= buttonPositions.cancelY + btnSize
-        ) {
-          this.cancelWarp();
-          return;
-        }
-
-        for (const node of warpState.nodes) {
-          const dx = logicalX - node.x;
-          const dy = logicalY - node.y;
-          if (dx * dx + dy * dy <= handleSize * handleSize) {
-            this.capturePointer(ev);
-            this.warp.startNodeDrag(node, logicalX, logicalY);
-            return;
-          }
-        }
-
-        return;
-      }
-
-      const puppetWarpState = this.puppetWarp.puppetWarpState();
-      if (puppetWarpState) {
-        const handleSize = Math.max(
-          3,
-          Math.round(5 / Math.max(0.001, this.scale())),
-        );
-
-        const buttonPositions = this.computePuppetWarpButtonPositions(
-          puppetWarpState,
-          this.scale(),
-        );
-        const btnSize = buttonPositions.btnSize;
-
-        if (
-          logicalX >= buttonPositions.commitX &&
-          logicalX <= buttonPositions.commitX + btnSize &&
-          logicalY >= buttonPositions.commitY &&
-          logicalY <= buttonPositions.commitY + btnSize
-        ) {
-          this.commitPuppetWarp();
-          return;
-        }
-
-        if (
-          logicalX >= buttonPositions.cancelX &&
-          logicalX <= buttonPositions.cancelX + btnSize &&
-          logicalY >= buttonPositions.cancelY &&
-          logicalY <= buttonPositions.cancelY + btnSize
-        ) {
-          this.cancelPuppetWarp();
-          return;
-        }
-
-        const clickedPin = this.puppetWarp.findPinAtPosition(
-          logicalX,
-          logicalY,
-          handleSize,
-        );
-        if (clickedPin) {
-          this.capturePointer(ev);
-          this.puppetWarp.startPinDrag(clickedPin, logicalX, logicalY);
-          this.puppetWarp.selectedPin.set(clickedPin);
-          return;
-        }
-
-        if (insideCanvas) {
-          this.puppetWarp.addPin(logicalX, logicalY);
-          return;
-        }
-
-        return;
-      }
-
-      const freeTransformState = this.freeTransform.transformState();
-      if (freeTransformState) {
-        const handleSize = Math.max(
-          3,
-          Math.round(5 / Math.max(0.001, this.scale())),
-        );
-        const positions = this.computeTransformButtonPositions(
-          freeTransformState,
-          this.scale(),
-        );
-        const buttonSize = positions.btnSize;
-
-        if (
-          logicalX >= positions.mirrorXX &&
-          logicalX <= positions.mirrorXX + buttonSize &&
-          logicalY >= positions.mirrorXY &&
-          logicalY <= positions.mirrorXY + buttonSize
-        ) {
-          this.freeTransformMirrorX = !this.freeTransformMirrorX;
-          this.renderLiveFreeTransformPreview();
-          return;
-        }
-
-        if (
-          logicalX >= positions.mirrorYX &&
-          logicalX <= positions.mirrorYX + buttonSize &&
-          logicalY >= positions.mirrorYY &&
-          logicalY <= positions.mirrorYY + buttonSize
-        ) {
-          this.freeTransformMirrorY = !this.freeTransformMirrorY;
-          this.renderLiveFreeTransformPreview();
-          return;
-        }
-
-        if (
-          logicalX >= positions.commitX &&
-          logicalX <= positions.commitX + buttonSize &&
-          logicalY >= positions.commitY &&
-          logicalY <= positions.commitY + buttonSize
-        ) {
-          this.commitFreeTransform();
-          return;
-        }
-
-        if (
-          logicalX >= positions.cancelX &&
-          logicalX <= positions.cancelX + buttonSize &&
-          logicalY >= positions.cancelY &&
-          logicalY <= positions.cancelY + buttonSize
-        ) {
-          this.cancelFreeTransform();
-          return;
-        }
-
-        const handles: TransformHandle[] = [
-          'top-left',
-          'top-center',
-          'top-right',
-          'middle-left',
-          'middle-right',
-          'bottom-left',
-          'bottom-center',
-          'bottom-right',
-          'rotate-center',
-        ];
-
-        for (const handle of handles) {
-          if (
-            this.freeTransform.isPointNearHandle(
-              logicalX,
-              logicalY,
-              handle,
-              freeTransformState,
-              handleSize,
-            )
-          ) {
-            this.capturePointer(ev);
-            this.freeTransform.startHandleDrag(handle, logicalX, logicalY);
-            return;
-          }
-        }
-
-        return;
-      }
-
-      const hasExistingSelection = !!this.document.selectionRect();
-      const clickedInSelection =
-        hasExistingSelection && this.isPointInSelection(logicalX, logicalY);
-      const isSelectTool =
-        tool === 'rect-select' ||
-        tool === 'ellipse-select' ||
-        tool === 'lasso-select';
-      if (
-        clickedInSelection &&
-        insideCanvas &&
-        this.moveSelectionHotkeyActive
-      ) {
-        const selectedLayer = this.document.selectedLayer();
-        if (selectedLayer?.locked) {
-          return;
-        }
-        this.document.beginMoveSelection('Move pixels');
-        this.capturePointer(ev);
-        this.selectionContentMoving = true;
-        this.selectionContentMoveStart = { x: logicalX, y: logicalY };
-        this.startSelectionContentMove();
-        return;
-      }
-      if (
-        clickedInSelection &&
-        insideCanvas &&
-        !ev.shiftKey &&
-        !ev.ctrlKey &&
-        isSelectTool
-      ) {
-        this.capturePointer(ev);
-        this.document.beginMoveSelection('Move selection');
-        this.selectionMoving = true;
-        this.selectionMoveStart = { x: logicalX, y: logicalY };
-        return;
-      }
-      if (isSelectTool && insideCanvas) {
-        this.capturePointer(ev);
-        if (tool === 'lasso-select') {
-          this.selectionStart = { x: logicalX, y: logicalY };
-          this.selectionDragging = true;
-          this.document.beginSelection(logicalX, logicalY, 'lasso' as any);
-          this.document.addLassoPoint(logicalX, logicalY);
-          return;
-        }
-
-        this.selectionStart = { x: logicalX, y: logicalY };
-        this.selectionDragging = true;
-        const shape = tool === 'ellipse-select' ? 'ellipse' : 'rect';
-        this.document.beginSelection(logicalX, logicalY, shape as any);
-        return;
-      }
-      if (
-        (tool === 'line' || tool === 'circle' || tool === 'square') &&
-        insideCanvas
-      ) {
-        const selectedLayer = this.document.selectedLayer();
-        if (selectedLayer?.locked) {
-          return;
-        }
-        this.capturePointer(ev);
-        if (tool === 'square' || tool === 'circle') {
-          this.shapeConstrainUniform.set(ev.shiftKey);
-        } else {
-          this.shapeConstrainUniform.set(false);
-        }
-        this.startShape(tool, logicalX, logicalY);
-        return;
-      }
-      if (tool === 'pen' && insideCanvas) {
-        const selectedLayer = this.document.selectedLayer();
-        if (selectedLayer?.locked) {
-          return;
-        }
-        this.capturePointer(ev);
-        this.addPenPoint(logicalX, logicalY);
-        return;
-      }
-      if (tool === 'eyedropper' && insideCanvas) {
-        const color = this.document.getColorAt(logicalX, logicalY);
-        if (color) {
-          this.tools.setEyedropperLastPickedColor(color);
-          if (this.colorPickerState.isPickingColor()) {
-            this.colorPickerState.deliverColor(color);
-          } else {
-            this.tools.setBrushColor(color);
-          }
-        }
-        return;
-      }
-      if (tool === 'fill' && insideCanvas) {
-        const selectedLayer = this.document.selectedLayer();
-        if (selectedLayer?.locked) {
-          return;
-        }
-        const selectionActive = this.document.selectionRect();
-        if (selectionActive && !this.isPointInSelection(logicalX, logicalY)) {
-          return;
-        }
-        this.document.saveSnapshot('Fill');
-        const layerId = this.document.selectedLayerId();
-        const fillMode = this.tools.fillMode();
-        if (fillMode === 'gradient') {
-          const gradientStartColor = this.tools.fillGradientStartColor();
-          const gradientEndColor = this.tools.fillGradientEndColor();
-          const gradientType = this.tools.fillGradientType();
-          const gradientAngle = this.tools.fillGradientAngle();
-          this.document.applyGradientFillToLayer(
-            layerId,
-            logicalX,
-            logicalY,
-            gradientStartColor,
-            gradientEndColor,
-            gradientType,
-            gradientAngle,
-          );
-        } else {
-          const fillColor = fillMode === 'erase' ? null : this.tools.fillColor();
-          this.document.applyFillToLayer(layerId, logicalX, logicalY, fillColor);
-        }
-      } else if (tool === 'bone' && insideCanvas) {
-        this.capturePointer(ev);
-        const frameId = this.getCurrentFrameId();
-        const clickedPoint = this.findBonePointAt(frameId, logicalX, logicalY);
-
-        if (clickedPoint) {
-          this.document.saveSnapshot('Move bone point');
-          this.draggingPointId = clickedPoint.pointId;
-          this.draggingPointBoneId = clickedPoint.boneId;
-          this.currentBoneId = clickedPoint.boneId;
-          this.boneService.selectPoint(clickedPoint.pointId);
-        } else {
-          this.document.saveSnapshot('Add bone point');
-          if (!this.currentBoneId) {
-            const newBone: Bone = {
-              id: `bone-${Date.now()}-${Math.random().toString(36).substring(7)}`,
-              points: [],
-              color: this.tools.boneColor(),
-              thickness: this.tools.boneThickness(),
-            };
-            this.currentBoneId = newBone.id;
-            this.boneService.addBone(frameId, newBone);
-          }
-
-          const newPoint: BonePoint = {
-            id: `point-${Date.now()}-${Math.random().toString(36).substring(7)}`,
-            x: logicalX,
-            y: logicalY,
-            parentId: undefined,
-          };
-
-          const bones = this.boneService.getBones(frameId);
-          const currentBone = bones.find((b) => b.id === this.currentBoneId);
-          if (currentBone && currentBone.points.length > 0) {
-            const selectedPoint = this.boneService.getSelectedPoint();
-            if (selectedPoint) {
-              newPoint.parentId = selectedPoint;
-            } else {
-              newPoint.parentId =
-                currentBone.points[currentBone.points.length - 1].id;
-            }
-          }
-
-          this.boneService.addPointToBone(
-            frameId,
-            this.currentBoneId,
-            newPoint,
-          );
-          this.boneService.selectPoint(newPoint.id);
-
-          const currentAnim = this.document.getCurrentAnimation();
-          if (currentAnim) {
-            const currentTime =
-              this.document.keyframeService?.getCurrentTime() || 0;
-            this.boneService.updatePoint(
-              frameId,
-              this.currentBoneId,
-              newPoint.id,
-              logicalX,
-              logicalY,
-              currentAnim.id,
-              currentTime,
-            );
-          }
-
-          if (this.tools.boneAutoBindEnabled()) {
-            const layerId = this.document.selectedLayerId();
-            const layerBuffer = this.document.getLayerBuffer(layerId);
-            if (layerBuffer) {
-              const radius = this.tools.boneAutoBindRadius();
-              this.boneService.autoBindPixels(
-                frameId,
-                layerBuffer,
-                w,
-                h,
-                this.currentBoneId,
-                newPoint.id,
-                logicalX,
-                logicalY,
-                radius,
-              );
-            }
-          }
-        }
-      } else if (tool === 'brush' || tool === 'eraser') {
-        const size =
-          tool === 'eraser' ? this.tools.eraserSize() : this.tools.brushSize();
-        const bSize = Math.max(1, size);
-        const half = Math.floor((bSize - 1) / 2);
-        const brushMinX = logicalX - half;
-        const brushMaxX = logicalX - half + bSize - 1;
-        const brushMinY = logicalY - half;
-        const brushMaxY = logicalY - half + bSize - 1;
-        const brushOverlapsDocument =
-          brushMaxX >= 0 && brushMinX < w && brushMaxY >= 0 && brushMinY < h;
-        if (brushOverlapsDocument) {
-          const selectedLayer = this.document.selectedLayer();
-          if (selectedLayer?.locked) {
-            return;
-          }
-          this.capturePointer(ev);
-          this.document.saveSnapshot('Paint');
-          this.painting = true;
-          this.lastPaintPos = { x: logicalX, y: logicalY };
-          const layerId = this.document.selectedLayerId();
-          const color = tool === 'eraser' ? null : this.tools.brushColor();
-          this.document.applyBrushToLayer(
-            layerId,
-            logicalX,
-            logicalY,
-            size,
-            color,
-            tool === 'eraser'
-              ? { eraserStrength: this.tools.eraserStrength() }
-              : undefined,
-          );
-        }
-      }
+  private releasePointer(ev: PointerEvent): void {
+    const target = ev.currentTarget as HTMLElement;
+    if (target?.releasePointerCapture) {
+      try { if (target.hasPointerCapture(ev.pointerId)) target.releasePointerCapture(ev.pointerId); } catch {}
     }
   }
 
@@ -1557,110 +635,15 @@ export class EditorCanvas implements OnDestroy {
     const currentPanY = this.panY();
     const logicalX = Math.floor((visX - currentPanX) / currentScale);
     const logicalY = Math.floor((visY - currentPanY) / currentScale);
-    const insideCanvas =
-      logicalX >= 0 && logicalX < w && logicalY >= 0 && logicalY < h;
-    if (!insideCanvas) {
-      this.closeContextMenu();
-      return;
-    }
-    const actions: ContextMenuAction[] = [];
-    const hasSelection = !!this.document.selectionRect();
-    const hasNonEmptySelection = this.hasNonEmptySelection();
-    if (hasSelection) {
-      actions.push({
-        id: 'generate',
-        labelKey: 'editor.canvas.menu.generate',
-        icon: 'heroIconsSparklesMini',
-        disabled: !hasNonEmptySelection,
-        submenu: [
-          {
-            id: 'generateFromLayer',
-            labelKey: 'pixelGeneration.generateFromLayer',
-            icon: 'heroIconsSquare2StackMini',
-          },
-          {
-            id: 'generateFromVisible',
-            labelKey: 'pixelGeneration.generateFromVisible',
-            icon: 'heroIconsEyeMini',
-          },
-        ],
-      });
-      actions.push({
-        id: 'invertSelection',
-        labelKey: 'editor.canvas.menu.invertSelection',
-        icon: 'heroIconsBarsArrowUpMini',
-        disabled: false,
-      });
-      actions.push({
-        id: 'growSelection',
-        labelKey: 'editor.canvas.menu.growSelection',
-        icon: 'heroIconsArrowsPointingOutMini',
-        disabled: false,
-        submenu: [
-          {
-            id: 'growBy1px',
-            labelKey: 'editor.canvas.menu.growBy1px',
-            icon: 'heroIconsPlusSmallMini',
-          },
-          {
-            id: 'growBy2px',
-            labelKey: 'editor.canvas.menu.growBy2px',
-            icon: 'heroIconsPlusSmallMini',
-          },
-          {
-            id: 'growBy5px',
-            labelKey: 'editor.canvas.menu.growBy5px',
-            icon: 'heroIconsPlusSmallMini',
-          },
-          {
-            id: 'growCustom',
-            labelKey: 'editor.canvas.menu.growCustom',
-            icon: 'heroIconsEllipsisHorizontalMini',
-          },
-        ],
-      });
-      actions.push({
-        id: 'makeCopyLayer',
-        labelKey: 'editor.canvas.menu.makeCopyLayer',
-        icon: 'heroIconsDocumentDuplicateMini',
-        disabled: !hasNonEmptySelection,
-      });
-      actions.push({
-        id: 'mergeVisibleToNewLayer',
-        labelKey: 'editor.canvas.menu.mergeVisibleToNewLayer',
-        icon: 'heroIconsRectangleStackMini',
-        disabled: false,
-      });
-      actions.push({
-        id: 'deselect',
-        labelKey: 'editor.canvas.menu.deselect',
-        icon: 'bootstrapBoundingBox',
-      });
-    }
-    if (!actions.length) {
-      this.closeContextMenu();
-      return;
-    }
-    const containerRect =
-      this.canvasContainer.nativeElement.getBoundingClientRect();
-    const offsetX = ev.clientX - containerRect.left;
-    const offsetY = ev.clientY - containerRect.top;
-    const estimatedWidth = 200;
-    const estimatedHeight = Math.max(40, actions.length * 36);
-    const maxX = Math.max(0, containerRect.width - estimatedWidth);
-    const maxY = Math.max(0, containerRect.height - estimatedHeight);
-    const clampedX = Math.max(0, Math.min(offsetX, maxX));
-    const clampedY = Math.max(0, Math.min(offsetY, maxY));
-    this.contextMenuPosition.set({ x: clampedX, y: clampedY });
-    this.contextMenuActions.set(actions);
-    this.contextMenuVisible.set(true);
+    const insideCanvas = logicalX >= 0 && logicalX < w && logicalY >= 0 && logicalY < h;
+    if (!insideCanvas) { this.closeContextMenu(); return; }
+    const containerRect = this.canvasContainer.nativeElement.getBoundingClientRect();
+    this.contextMenu.show(ev.clientX, ev.clientY, containerRect, this.hasNonEmptySelection());
   }
 
   private hasNonEmptySelection(): boolean {
     const sel = this.document.selectionRect();
     if (!sel) return false;
-    const shape = this.document.selectionShape();
-    const poly = this.document.selectionPolygon();
     const layerId = this.document.selectedLayerId();
     const buf = this.document.getLayerBuffer(layerId);
     if (!buf) return false;
@@ -1668,187 +651,37 @@ export class EditorCanvas implements OnDestroy {
     const h = this.document.canvasHeight();
     for (let y = sel.y; y < sel.y + sel.height && y < h; y++) {
       for (let x = sel.x; x < sel.x + sel.width && x < w; x++) {
-        if (
-          this.isPointInSelection(x, y) &&
-          buf[y * w + x] &&
-          buf[y * w + x].length > 0
-        ) {
-          return true;
-        }
+        if (this.isPointInSelection(x, y) && buf[y * w + x] && buf[y * w + x].length > 0) return true;
       }
     }
     return false;
   }
 
-  closeContextMenu() {
-    this.contextMenuVisible.set(false);
-    this.contextMenuActions.set([]);
-    this.submenuVisible.set(false);
-    this.submenuActions.set([]);
-  }
+  closeContextMenu() { this.contextMenu.close(); }
 
-  onSubmenuTrigger(
-    action: ContextMenuAction,
-    event: MouseEvent,
-    buttonElement: HTMLElement,
-  ) {
+  onSubmenuTrigger(action: ContextMenuAction, event: MouseEvent, buttonElement: HTMLElement) {
     if (!action.submenu || action.submenu.length === 0) return;
     event.stopPropagation();
     const rect = buttonElement.getBoundingClientRect();
-    const containerRect =
-      this.canvasContainer.nativeElement.getBoundingClientRect();
-    const submenuX = rect.right - containerRect.left + 4;
-    const submenuY = rect.top - containerRect.top;
-    this.submenuPosition.set({ x: submenuX, y: submenuY });
-    this.submenuActions.set(action.submenu);
-    this.submenuVisible.set(true);
+    const containerRect = this.canvasContainer.nativeElement.getBoundingClientRect();
+    this.contextMenu.showSubmenu(action, rect, containerRect);
   }
 
   onContextMenuAction(actionId: ContextMenuActionId, event?: MouseEvent) {
-    if (actionId === 'generateFromLayer') {
-      this.handleGenerateFromSelection('layer');
-    } else if (actionId === 'generateFromVisible') {
-      this.handleGenerateFromSelection('visible');
-    } else if (actionId === 'deselect') {
-      this.document.clearSelection();
-    } else if (actionId === 'invertSelection') {
-      this.document.invertSelection();
-    } else if (actionId === 'growBy1px') {
-      this.document.growSelection(1);
-    } else if (actionId === 'growBy2px') {
-      this.document.growSelection(2);
-    } else if (actionId === 'growBy5px') {
-      this.document.growSelection(5);
-    } else if (actionId === 'growCustom') {
-      if (event) {
-        const containerRect =
-          this.canvasContainer.nativeElement.getBoundingClientRect();
-        const offsetX = event.clientX - containerRect.left;
-        const offsetY = event.clientY - containerRect.top;
-        this.inputDialogPosition.set({ x: offsetX + 10, y: offsetY });
-        this.inputDialogTitle.set('Enter growth amount (pixels):');
-        this.inputDialogValue.set('10');
-        this.inputDialogCallback.set((value: string) => {
-          const parsed = parseInt(value, 10);
-          if (!isNaN(parsed) && parsed > 0) {
-            this.document.growSelection(parsed);
-          }
-          this.closeInputDialog();
-        });
-        this.inputDialogVisible.set(true);
-        return;
-      }
-    } else if (actionId === 'makeCopyLayer') {
-      this.document.makeCopyLayer();
-    } else if (actionId === 'mergeVisibleToNewLayer') {
-      this.document.mergeVisibleToNewLayer();
-    }
-    this.closeContextMenu();
+    const containerRect = this.canvasContainer.nativeElement.getBoundingClientRect();
+    this.contextMenu.executeAction(
+      actionId,
+      event,
+      containerRect,
+      () => this.handleGenerateFromSelection('layer'),
+      () => this.handleGenerateFromSelection('visible'),
+    );
   }
 
-  closeInputDialog() {
-    this.inputDialogVisible.set(false);
-    this.inputDialogCallback.set(null);
-  }
-
-  getShortcutForAction(actionId: ContextMenuActionId): string | null {
-    const hotkeyMap: Record<ContextMenuActionId, string | null> = {
-      deselect: 'edit.deselect',
-      invertSelection: 'edit.invertSelection',
-      growSelection: null,
-      growBy1px: 'edit.growBy1px',
-      growBy2px: 'edit.growBy2px',
-      growBy5px: 'edit.growBy5px',
-      growCustom: null,
-      makeCopyLayer: 'edit.makeCopyLayer',
-      mergeVisibleToNewLayer: 'edit.mergeVisibleToNewLayer',
-      generate: null,
-      generateFromLayer: null,
-      generateFromVisible: null,
-    };
-    const hotkeyId = hotkeyMap[actionId];
-    if (!hotkeyId) return null;
-    const binding = this.hotkeys.getBinding(hotkeyId);
-    return binding ? this.hotkeys.keyStringToDisplay(binding) : null;
-  }
-
-  onInputDialogSubmit() {
-    const callback = this.inputDialogCallback();
-    if (callback) {
-      callback(this.inputDialogValue());
-    }
-  }
-
-  onInputDialogCancel() {
-    this.closeInputDialog();
-  }
-
-  onPointerUp(ev: PointerEvent) {
-    this.releasePointer(ev);
-    this.panning = false;
-
-    if (this.freeTransform.isDraggingHandle()) {
-      this.freeTransform.endHandleDrag();
-      return;
-    }
-
-    if (this.distort.isDraggingHandle()) {
-      this.distort.endHandleDrag();
-      return;
-    }
-
-    if (this.perspective.isDraggingHandle()) {
-      this.perspective.endHandleDrag();
-      return;
-    }
-
-    if (this.warp.isDraggingNode()) {
-      this.warp.endNodeDrag();
-      return;
-    }
-
-    if (this.puppetWarp.isDraggingPin()) {
-      this.puppetWarp.endPinDrag();
-      return;
-    }
-
-    if (this.shaping) {
-      this.finishShape(ev.shiftKey);
-    }
-    if (this.painting) {
-      this.painting = false;
-      this.lastPaintPos = null;
-    }
-
-    if (this.draggingPointId) {
-      this.draggingPointId = null;
-      this.draggingPointBoneId = null;
-      return;
-    }
-
-    if (this.selectionMoving) {
-      this.selectionMoving = false;
-      this.selectionMoveStart = null;
-      this.document.endMoveSelection('Move selection');
-      return;
-    }
-
-    if (this.selectionContentMoving) {
-      this.selectionContentMoving = false;
-      this.selectionContentMoveStart = null;
-      this.endSelectionContentMove();
-      this.document.endMoveSelection('Move pixels');
-      return;
-    }
-
-    if (this.selectionDragging) {
-      this.selectionDragging = false;
-      this.selectionStart = null;
-      this.document.endSelection();
-    }
-  }
-
-  infoVisible = signal(true);
+  closeInputDialog() { this.contextMenu.closeInputDialog(); }
+  getShortcutForAction(actionId: ContextMenuActionId): string | null { return this.contextMenu.getShortcutForAction(actionId); }
+  onInputDialogSubmit() { this.contextMenu.submitInputDialog(); }
+  onInputDialogCancel() { this.closeInputDialog(); }
 
   startSelectionContentMove() {
     const sel = this.document.selectionRect();
@@ -1859,12 +692,7 @@ export class EditorCanvas implements OnDestroy {
     const w = this.document.canvasWidth();
     const h = this.document.canvasHeight();
     this.originalLayerId = layerId;
-    this.movingContentOriginalRect = {
-      x: sel.x,
-      y: sel.y,
-      width: sel.width,
-      height: sel.height,
-    };
+    this.movingContentOriginalRect = { x: sel.x, y: sel.y, width: sel.width, height: sel.height };
     this.movingContentBuffer = [];
     for (let y = 0; y < h; y++) {
       for (let x = 0; x < w; x++) {
@@ -1880,17 +708,10 @@ export class EditorCanvas implements OnDestroy {
     this.document.layerPixelsVersion.update((v) => v + 1);
   }
 
-  moveSelectionContent(dx: number, dy: number) {
-    this.document.moveSelection(dx, dy);
-  }
+  moveSelectionContent(dx: number, dy: number) { this.document.moveSelection(dx, dy); }
 
   endSelectionContentMove() {
-    if (
-      !this.movingContentBuffer ||
-      !this.originalLayerId ||
-      !this.movingContentOriginalRect
-    )
-      return;
+    if (!this.movingContentBuffer || !this.originalLayerId || !this.movingContentOriginalRect) return;
     const originalBuf = this.document.getLayerBuffer(this.originalLayerId);
     if (!originalBuf) return;
     const w = this.document.canvasWidth();
@@ -1924,13 +745,7 @@ export class EditorCanvas implements OnDestroy {
     if (width > 0) {
       this.document.setCanvasSize(width, this.document.canvasHeight());
       const flatLayers = this.document.getFlattenedLayers();
-      for (const l of flatLayers) {
-        this.document.ensureLayerBuffer(
-          l.id,
-          width,
-          this.document.canvasHeight(),
-        );
-      }
+      for (const l of flatLayers) this.document.ensureLayerBuffer(l.id, width, this.document.canvasHeight());
     }
   }
 
@@ -1940,33 +755,11 @@ export class EditorCanvas implements OnDestroy {
     if (height > 0) {
       this.document.setCanvasSize(this.document.canvasWidth(), height);
       const flatLayers = this.document.getFlattenedLayers();
-      for (const l of flatLayers) {
-        this.document.ensureLayerBuffer(
-          l.id,
-          this.document.canvasWidth(),
-          height,
-        );
-      }
+      for (const l of flatLayers) this.document.ensureLayerBuffer(l.id, this.document.canvasWidth(), height);
     }
   }
 
-  // Note: per-layer pixel buffers are stored in EditorStateService; ensureLayerBuffer
-  // calls that service method when needed.
-
-  // applyBrush removed: logic delegated to EditorDocumentService.applyBrushToLayer
-
-  // Draw a straight line between two logical pixel coordinates and apply the
-  // brush/eraser at each step so fast pointer movement produces continuous
-  // strokes. Uses Bresenham's line algorithm for integer rasterization.
-  private drawLinePaint(
-    layerId: string,
-    x0: number,
-    y0: number,
-    x1: number,
-    y1: number,
-    brushSize: number,
-    color: string | null,
-  ) {
+  private drawLinePaint(layerId: string, x0: number, y0: number, x1: number, y1: number, brushSize: number, color: string | null) {
     const dx = Math.abs(x1 - x0);
     const sx = x0 < x1 ? 1 : -1;
     const dy = -Math.abs(y1 - y0);
@@ -1974,29 +767,13 @@ export class EditorCanvas implements OnDestroy {
     let err = dx + dy;
     let x = x0;
     let y = y0;
-    const eraserOptions =
-      color === null
-        ? { eraserStrength: this.tools.eraserStrength() }
-        : undefined;
+    const eraserOptions = color === null ? { eraserStrength: this.tools.eraserStrength() } : undefined;
     while (true) {
-      this.document.applyBrushToLayer(
-        layerId,
-        x,
-        y,
-        brushSize,
-        color,
-        eraserOptions,
-      );
+      this.document.applyBrushToLayer(layerId, x, y, brushSize, color, eraserOptions);
       if (x === x1 && y === y1) break;
       const e2 = 2 * err;
-      if (e2 >= dy) {
-        err += dy;
-        x += sx;
-      }
-      if (e2 <= dx) {
-        err += dx;
-        y += sy;
-      }
+      if (e2 >= dy) { err += dy; x += sx; }
+      if (e2 <= dx) { err += dx; y += sy; }
     }
   }
 
@@ -2007,1225 +784,27 @@ export class EditorCanvas implements OnDestroy {
     this.applyZoom(v);
   }
 
-  resetRotation() {
-    this.rotation.set(0);
-  }
-
-  increaseZoom(step = 0.1) {
-    const factor = 1 + Math.max(0, step);
-    this.applyZoom(this.scale() * factor);
-  }
-
-  decreaseZoom(step = 0.1) {
-    const factor = 1 + Math.max(0, step);
-    this.applyZoom(this.scale() / factor);
-  }
-
-  private activateFreeTransform(): void {
-    const sel = this.document.selectionRect();
-    if (!sel || sel.width <= 0 || sel.height <= 0) {
-      return;
-    }
-    this.freeTransformDuplicate = false;
-    this.freeTransformPreviewIndices = null;
-    this.freeTransformPreviewBackup = null;
-    // Pre-change snapshot so a single undo reverts the whole transform session.
-    this.document.saveSnapshot('Free Transform');
-    const layer = this.document.selectedLayer();
-    if (!layer || !isLayer(layer)) return;
-    const layerBuf = this.document.getLayerBuffer(layer.id);
-    if (!layerBuf) return;
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-    const original: string[] = new Array<string>(sel.width * sel.height).fill(
-      '',
-    );
-    for (let y = 0; y < sel.height; y++) {
-      for (let x = 0; x < sel.width; x++) {
-        const srcX = sel.x + x;
-        const srcY = sel.y + y;
-        if (srcX >= 0 && srcX < canvasW && srcY >= 0 && srcY < canvasH) {
-          const srcIdx = srcY * canvasW + srcX;
-          const dstIdx = y * sel.width + x;
-          original[dstIdx] = layerBuf[srcIdx] || '';
-          layerBuf[srcIdx] = '';
-        }
-      }
-    }
-    this.document.layerPixelsVersion.update((v) => v + 1);
-    this.freeTransformOriginalBuffer = original;
-    this.freeTransformOriginalRect = {
-      x: sel.x,
-      y: sel.y,
-      width: sel.width,
-      height: sel.height,
-    };
-    this.freeTransformLayerId = layer.id;
-    this.freeTransform.startTransform(sel.x, sel.y, sel.width, sel.height);
-    // Render initial preview into layer so user sees live pixels immediately
-    this.renderLiveFreeTransformPreview();
-  }
-
-  private commitFreeTransform(): void {
-    const state = this.freeTransform.commitTransform();
-    if (!state) return;
-    const sel = this.document.selectionRect();
-    if (!sel) return;
-
-    const scaleX = state.width / sel.width;
-    const scaleY = state.height / sel.height;
-    const doScale =
-      Math.abs(scaleX - 1) > 0.0001 || Math.abs(scaleY - 1) > 0.0001;
-    const doRotate = Math.abs(state.rotation) > 0.0001;
-
-    // Restore pixels from backup first to avoid losing content outside selection
-    if (
-      this.freeTransformPreviewIndices &&
-      this.freeTransformPreviewIndices.length &&
-      this.freeTransformLayerId &&
-      this.freeTransformPreviewBackup
-    ) {
-      const layerBuffer = this.document.getLayerBuffer(
-        this.freeTransformLayerId,
-      );
-      if (layerBuffer) {
-        for (const idx of this.freeTransformPreviewIndices) {
-          if (idx >= 0 && idx < layerBuffer.length) {
-            const backupValue = this.freeTransformPreviewBackup.get(idx);
-            layerBuffer[idx] = backupValue !== undefined ? backupValue : '';
-          }
-        }
-      }
-    }
-    if (
-      this.freeTransformDuplicate &&
-      this.freeTransformOriginalBuffer &&
-      this.freeTransformOriginalRect &&
-      this.freeTransformLayerId
-    ) {
-      this.restoreOriginalTransformPixels();
-      this.applyTransformToSelection(
-        state,
-        scaleX,
-        scaleY,
-        this.freeTransformOriginalBuffer,
-        this.freeTransformOriginalRect,
-      );
-    } else {
-      if (doScale || doRotate) {
-        if (
-          this.freeTransformOriginalBuffer &&
-          this.freeTransformOriginalRect &&
-          this.freeTransformLayerId
-        ) {
-          this.applyTransformToSelection(
-            state,
-            scaleX,
-            scaleY,
-            this.freeTransformOriginalBuffer,
-            this.freeTransformOriginalRect,
-          );
-        }
-      } else {
-        this.restoreOriginalTransformPixels();
-      }
-    }
-
-    this.document.updateSelectionBounds(
-      Math.floor(state.x),
-      Math.floor(state.y),
-      Math.floor(state.width),
-      Math.floor(state.height),
-    );
-    // Mark canvas dirty after transform
-    this.document.setCanvasSaved(false);
-    this.freeTransformOriginalBuffer = null;
-    this.freeTransformOriginalRect = null;
-    this.freeTransformLayerId = null;
-    this.freeTransformDuplicate = false;
-    this.freeTransformPreviewIndices = null;
-    this.freeTransformPreviewBackup = null;
-  }
-
-  private cancelFreeTransform(): void {
-    // Restore pixels from backup before clearing preview
-    if (
-      this.freeTransformPreviewIndices &&
-      this.freeTransformLayerId &&
-      this.freeTransformPreviewBackup
-    ) {
-      const buf = this.document.getLayerBuffer(this.freeTransformLayerId);
-      if (buf) {
-        for (const idx of this.freeTransformPreviewIndices) {
-          if (idx >= 0 && idx < buf.length) {
-            const backupValue = this.freeTransformPreviewBackup.get(idx);
-            buf[idx] = backupValue !== undefined ? backupValue : '';
-          }
-        }
-        this.document.layerPixelsVersion.update((v) => v + 1);
-      }
-    }
-    this.restoreOriginalTransformPixels();
-    this.freeTransform.cancelTransform();
-    this.freeTransformOriginalBuffer = null;
-    this.freeTransformOriginalRect = null;
-    this.freeTransformLayerId = null;
-    this.freeTransformDuplicate = false;
-    this.freeTransformPreviewIndices = null;
-    this.freeTransformPreviewBackup = null;
-  }
-
-  private activateDistort(): void {
-    const sel = this.document.selectionRect();
-    if (!sel || sel.width <= 0 || sel.height <= 0) {
-      return;
-    }
-
-    this.document.saveSnapshot('Distort');
-    const layer = this.document.selectedLayer();
-    if (!layer || !isLayer(layer)) return;
-    const layerBuf = this.document.getLayerBuffer(layer.id);
-    if (!layerBuf) return;
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    this.distortFullLayerBackup = [...layerBuf];
-
-    const original: string[] = new Array<string>(sel.width * sel.height).fill(
-      '',
-    );
-    for (let y = 0; y < sel.height; y++) {
-      for (let x = 0; x < sel.width; x++) {
-        const srcX = sel.x + x;
-        const srcY = sel.y + y;
-        if (srcX >= 0 && srcX < canvasW && srcY >= 0 && srcY < canvasH) {
-          const srcIdx = srcY * canvasW + srcX;
-          const dstIdx = y * sel.width + x;
-          original[dstIdx] = layerBuf[srcIdx] || '';
-          layerBuf[srcIdx] = '';
-        }
-      }
-    }
-    this.document.layerPixelsVersion.update((v) => v + 1);
-    this.distortOriginalBuffer = original;
-    this.distortOriginalRect = {
-      x: sel.x,
-      y: sel.y,
-      width: sel.width,
-      height: sel.height,
-    };
-    this.distortLayerId = layer.id;
-    this.renderLiveDistortPreview();
-  }
-
-  private activatePerspective(): void {
-    const sel = this.document.selectionRect();
-    if (!sel || sel.width <= 0 || sel.height <= 0) {
-      return;
-    }
-
-    this.document.saveSnapshot('Perspective');
-    const layer = this.document.selectedLayer();
-    if (!layer || !isLayer(layer)) return;
-    const layerBuf = this.document.getLayerBuffer(layer.id);
-    if (!layerBuf) return;
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    this.perspectiveFullLayerBackup = [...layerBuf];
-
-    const original: string[] = new Array<string>(sel.width * sel.height).fill(
-      '',
-    );
-    for (let y = 0; y < sel.height; y++) {
-      for (let x = 0; x < sel.width; x++) {
-        const srcX = sel.x + x;
-        const srcY = sel.y + y;
-        if (srcX >= 0 && srcX < canvasW && srcY >= 0 && srcY < canvasH) {
-          const srcIdx = srcY * canvasW + srcX;
-          const dstIdx = y * sel.width + x;
-          original[dstIdx] = layerBuf[srcIdx] || '';
-          layerBuf[srcIdx] = '';
-        }
-      }
-    }
-    this.document.layerPixelsVersion.update((v) => v + 1);
-    this.perspectiveOriginalBuffer = original;
-    this.perspectiveOriginalRect = {
-      x: sel.x,
-      y: sel.y,
-      width: sel.width,
-      height: sel.height,
-    };
-    this.perspectiveLayerId = layer.id;
-    this.renderLivePerspectivePreview();
-  }
-
-  private commitDistort(): void {
-    const state = this.distort.commitDistort();
-    if (!state) return;
-
-    if (
-      !this.distortOriginalBuffer ||
-      !this.distortOriginalRect ||
-      !this.distortLayerId ||
-      !this.distortFullLayerBackup
-    ) {
-      return;
-    }
-
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    const originalLayerBuffer = this.document.getLayerBuffer(
-      this.distortLayerId,
-    );
-    if (!originalLayerBuffer) return;
-
-    for (let i = 0; i < originalLayerBuffer.length; i++) {
-      originalLayerBuffer[i] = this.distortFullLayerBackup[i];
-    }
-
-    for (let y = 0; y < this.distortOriginalRect.height; y++) {
-      for (let x = 0; x < this.distortOriginalRect.width; x++) {
-        const destX: number = this.distortOriginalRect.x + x;
-        const destY: number = this.distortOriginalRect.y + y;
-        if (destX >= 0 && destX < canvasW && destY >= 0 && destY < canvasH) {
-          const destIdx = destY * canvasW + destX;
-          originalLayerBuffer[destIdx] = '';
-        }
-      }
-    }
-
-    const originalSelectionLayer = this.document.addLayer('Original Selection');
-    const origSelLayerBuf = this.document.getLayerBuffer(
-      originalSelectionLayer.id,
-    );
-
-    if (origSelLayerBuf) {
-      for (let y = 0; y < this.distortOriginalRect.height; y++) {
-        for (let x = 0; x < this.distortOriginalRect.width; x++) {
-          const srcIdx = y * this.distortOriginalRect.width + x;
-          const destX: number = this.distortOriginalRect.x + x;
-          const destY: number = this.distortOriginalRect.y + y;
-          if (destX >= 0 && destX < canvasW && destY >= 0 && destY < canvasH) {
-            const destIdx = destY * canvasW + destX;
-            origSelLayerBuf[destIdx] = this.distortOriginalBuffer[srcIdx] || '';
-          }
-        }
-      }
-    }
-
-    this.document.toggleLayerVisibility(originalSelectionLayer.id);
-
-    const srcCorners = [
-      { x: 0, y: 0 },
-      { x: state.sourceWidth, y: 0 },
-      { x: state.sourceWidth, y: state.sourceHeight },
-      { x: 0, y: state.sourceHeight },
-    ];
-
-    const dstCorners = [
-      state.corners.topLeft,
-      state.corners.topRight,
-      state.corners.bottomRight,
-      state.corners.bottomLeft,
-    ];
-
-    const result = this.transformService.applyDistort(
-      this.distortOriginalBuffer,
-      this.distortOriginalRect.width,
-      this.distortOriginalRect.height,
-      srcCorners,
-      dstCorners,
-    );
-
-    const distortedLayer = this.document.addLayer('Distorted Layer');
-    const distortedLayerBuf = this.document.getLayerBuffer(distortedLayer.id);
-
-    if (distortedLayerBuf) {
-      for (let y = 0; y < result.height; y++) {
-        for (let x = 0; x < result.width; x++) {
-          const srcIdx = y * result.width + x;
-          const color = result.buffer[srcIdx];
-          if (color) {
-            const destX = Math.floor(result.minX + x);
-            const destY = Math.floor(result.minY + y);
-            if (
-              destX >= 0 &&
-              destX < canvasW &&
-              destY >= 0 &&
-              destY < canvasH
-            ) {
-              const destIdx = destY * canvasW + destX;
-              distortedLayerBuf[destIdx] = color;
-            }
-          }
-        }
-      }
-    }
-
-    this.document.selectLayer(distortedLayer.id);
-
-    this.document.layerPixelsVersion.update((v) => v + 1);
-    this.document.setCanvasSaved(false);
-    this.distortOriginalBuffer = null;
-    this.distortOriginalRect = null;
-    this.distortLayerId = null;
-    this.distortFullLayerBackup = null;
-  }
-
-  private cancelDistort(): void {
-    if (this.distortFullLayerBackup && this.distortLayerId) {
-      const layerBuffer = this.document.getLayerBuffer(this.distortLayerId);
-      if (layerBuffer) {
-        for (let i = 0; i < layerBuffer.length; i++) {
-          layerBuffer[i] = this.distortFullLayerBackup[i];
-        }
-        this.document.layerPixelsVersion.update((v) => v + 1);
-      }
-    }
-
-    this.distort.cancelDistort();
-    this.distortOriginalBuffer = null;
-    this.distortOriginalRect = null;
-    this.distortLayerId = null;
-    this.distortFullLayerBackup = null;
-  }
-
-  private renderLiveDistortPreview(): void {
-    const state = this.distort.distortState();
-    if (
-      !state ||
-      !this.distortOriginalBuffer ||
-      !this.distortOriginalRect ||
-      !this.distortLayerId ||
-      !this.distortFullLayerBackup
-    ) {
-      return;
-    }
-
-    const layerBuffer = this.document.getLayerBuffer(this.distortLayerId);
-    if (!layerBuffer) return;
-
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    for (let i = 0; i < layerBuffer.length; i++) {
-      layerBuffer[i] = this.distortFullLayerBackup[i];
-    }
-
-    const srcCorners = [
-      { x: 0, y: 0 },
-      { x: state.sourceWidth, y: 0 },
-      { x: state.sourceWidth, y: state.sourceHeight },
-      { x: 0, y: state.sourceHeight },
-    ];
-
-    const dstCorners = [
-      state.corners.topLeft,
-      state.corners.topRight,
-      state.corners.bottomRight,
-      state.corners.bottomLeft,
-    ];
-
-    const result = this.transformService.applyDistort(
-      this.distortOriginalBuffer,
-      this.distortOriginalRect.width,
-      this.distortOriginalRect.height,
-      srcCorners,
-      dstCorners,
-    );
-
-    for (let y = 0; y < result.height; y++) {
-      for (let x = 0; x < result.width; x++) {
-        const srcIdx = y * result.width + x;
-        const color = result.buffer[srcIdx];
-        if (color) {
-          const destX = Math.floor(result.minX + x);
-          const destY = Math.floor(result.minY + y);
-          if (destX >= 0 && destX < canvasW && destY >= 0 && destY < canvasH) {
-            const destIdx = destY * canvasW + destX;
-            layerBuffer[destIdx] = color;
-          }
-        }
-      }
-    }
-
-    this.document.layerPixelsVersion.update((v) => v + 1);
-  }
-
-  private commitPerspective(): void {
-    const state = this.perspective.commitPerspective();
-    if (!state) return;
-
-    if (
-      !this.perspectiveOriginalBuffer ||
-      !this.perspectiveOriginalRect ||
-      !this.perspectiveLayerId ||
-      !this.perspectiveFullLayerBackup
-    ) {
-      return;
-    }
-
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    const originalLayerBuffer = this.document.getLayerBuffer(
-      this.perspectiveLayerId,
-    );
-    if (!originalLayerBuffer) return;
-
-    for (let i = 0; i < originalLayerBuffer.length; i++) {
-      originalLayerBuffer[i] = this.perspectiveFullLayerBackup[i];
-    }
-
-    for (let y = 0; y < this.perspectiveOriginalRect.height; y++) {
-      for (let x = 0; x < this.perspectiveOriginalRect.width; x++) {
-        const destX: number = this.perspectiveOriginalRect.x + x;
-        const destY: number = this.perspectiveOriginalRect.y + y;
-        if (destX >= 0 && destX < canvasW && destY >= 0 && destY < canvasH) {
-          const destIdx = destY * canvasW + destX;
-          originalLayerBuffer[destIdx] = '';
-        }
-      }
-    }
-
-    const originalSelectionLayer = this.document.addLayer('Original Selection');
-    const origSelLayerBuf = this.document.getLayerBuffer(
-      originalSelectionLayer.id,
-    );
-
-    if (origSelLayerBuf) {
-      for (let y = 0; y < this.perspectiveOriginalRect.height; y++) {
-        for (let x = 0; x < this.perspectiveOriginalRect.width; x++) {
-          const srcIdx = y * this.perspectiveOriginalRect.width + x;
-          const destX: number = this.perspectiveOriginalRect.x + x;
-          const destY: number = this.perspectiveOriginalRect.y + y;
-          if (destX >= 0 && destX < canvasW && destY >= 0 && destY < canvasH) {
-            const destIdx = destY * canvasW + destX;
-            origSelLayerBuf[destIdx] =
-              this.perspectiveOriginalBuffer[srcIdx] || '';
-          }
-        }
-      }
-    }
-
-    this.document.toggleLayerVisibility(originalSelectionLayer.id);
-
-    const srcCorners = [
-      { x: 0, y: 0 },
-      { x: state.sourceWidth, y: 0 },
-      { x: state.sourceWidth, y: state.sourceHeight },
-      { x: 0, y: state.sourceHeight },
-    ];
-
-    const dstCorners = [
-      state.corners.topLeft,
-      state.corners.topRight,
-      state.corners.bottomRight,
-      state.corners.bottomLeft,
-    ];
-
-    const result = this.transformService.applyDistort(
-      this.perspectiveOriginalBuffer,
-      this.perspectiveOriginalRect.width,
-      this.perspectiveOriginalRect.height,
-      srcCorners,
-      dstCorners,
-    );
-
-    const perspectiveLayer = this.document.addLayer('Perspective Layer');
-    const perspectiveLayerBuf = this.document.getLayerBuffer(
-      perspectiveLayer.id,
-    );
-
-    if (perspectiveLayerBuf) {
-      for (let y = 0; y < result.height; y++) {
-        for (let x = 0; x < result.width; x++) {
-          const srcIdx = y * result.width + x;
-          const color = result.buffer[srcIdx];
-          if (color) {
-            const destX = Math.floor(result.minX + x);
-            const destY = Math.floor(result.minY + y);
-            if (
-              destX >= 0 &&
-              destX < canvasW &&
-              destY >= 0 &&
-              destY < canvasH
-            ) {
-              const destIdx = destY * canvasW + destX;
-              perspectiveLayerBuf[destIdx] = color;
-            }
-          }
-        }
-      }
-    }
-
-    this.document.selectLayer(perspectiveLayer.id);
-
-    this.document.layerPixelsVersion.update((v) => v + 1);
-    this.document.setCanvasSaved(false);
-    this.perspectiveOriginalBuffer = null;
-    this.perspectiveOriginalRect = null;
-    this.perspectiveLayerId = null;
-    this.perspectiveFullLayerBackup = null;
-  }
-
-  private cancelPerspective(): void {
-    if (this.perspectiveFullLayerBackup && this.perspectiveLayerId) {
-      const layerBuffer = this.document.getLayerBuffer(this.perspectiveLayerId);
-      if (layerBuffer) {
-        for (let i = 0; i < layerBuffer.length; i++) {
-          layerBuffer[i] = this.perspectiveFullLayerBackup[i];
-        }
-        this.document.layerPixelsVersion.update((v) => v + 1);
-      }
-    }
-
-    this.perspective.cancelPerspective();
-    this.perspectiveOriginalBuffer = null;
-    this.perspectiveOriginalRect = null;
-    this.perspectiveLayerId = null;
-    this.perspectiveFullLayerBackup = null;
-  }
-
-  private activateWarp(): void {
-    const sel = this.document.selectionRect();
-    if (!sel || sel.width <= 0 || sel.height <= 0) {
-      return;
-    }
-
-    this.document.saveSnapshot('Warp');
-    const layer = this.document.selectedLayer();
-    if (!layer || !isLayer(layer)) return;
-    const layerBuf = this.document.getLayerBuffer(layer.id);
-    if (!layerBuf) return;
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    this.warpFullLayerBackup = [...layerBuf];
-
-    const original: string[] = new Array<string>(sel.width * sel.height).fill(
-      '',
-    );
-    for (let y = 0; y < sel.height; y++) {
-      for (let x = 0; x < sel.width; x++) {
-        const srcX = sel.x + x;
-        const srcY = sel.y + y;
-        if (srcX >= 0 && srcX < canvasW && srcY >= 0 && srcY < canvasH) {
-          const srcIdx = srcY * canvasW + srcX;
-          const dstIdx = y * sel.width + x;
-          original[dstIdx] = layerBuf[srcIdx] || '';
-          layerBuf[srcIdx] = '';
-        }
-      }
-    }
-    this.document.layerPixelsVersion.update((v) => v + 1);
-    this.warpOriginalBuffer = original;
-    this.warpOriginalRect = {
-      x: sel.x,
-      y: sel.y,
-      width: sel.width,
-      height: sel.height,
-    };
-    this.warpLayerId = layer.id;
-    this.renderLiveWarpPreview();
-  }
-
-  private activatePuppetWarp(): void {
-    const sel = this.document.selectionRect();
-    if (!sel || sel.width <= 0 || sel.height <= 0) {
-      return;
-    }
-
-    this.document.saveSnapshot('Puppet Warp');
-    const layer = this.document.selectedLayer();
-    if (!layer || !isLayer(layer)) return;
-    const layerBuf = this.document.getLayerBuffer(layer.id);
-    if (!layerBuf) return;
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    this.puppetWarpFullLayerBackup = [...layerBuf];
-
-    const original: string[] = new Array<string>(sel.width * sel.height).fill(
-      '',
-    );
-    for (let y = 0; y < sel.height; y++) {
-      for (let x = 0; x < sel.width; x++) {
-        const srcX = sel.x + x;
-        const srcY = sel.y + y;
-        if (srcX >= 0 && srcX < canvasW && srcY >= 0 && srcY < canvasH) {
-          const srcIdx = srcY * canvasW + srcX;
-          const dstIdx = y * sel.width + x;
-          original[dstIdx] = layerBuf[srcIdx] || '';
-          layerBuf[srcIdx] = '';
-        }
-      }
-    }
-    this.document.layerPixelsVersion.update((v) => v + 1);
-    this.puppetWarpOriginalBuffer = original;
-    this.puppetWarpOriginalRect = {
-      x: sel.x,
-      y: sel.y,
-      width: sel.width,
-      height: sel.height,
-    };
-    this.puppetWarpLayerId = layer.id;
-    this.renderLivePuppetWarpPreview();
-  }
-
-  private commitWarp(): void {
-    const state = this.warp.commitWarp();
-    if (!state) return;
-
-    if (
-      !this.warpOriginalBuffer ||
-      !this.warpOriginalRect ||
-      !this.warpLayerId ||
-      !this.warpFullLayerBackup
-    ) {
-      return;
-    }
-
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    const originalLayerBuffer = this.document.getLayerBuffer(this.warpLayerId);
-    if (!originalLayerBuffer) return;
-
-    for (let i = 0; i < originalLayerBuffer.length; i++) {
-      originalLayerBuffer[i] = this.warpFullLayerBackup[i];
-    }
-
-    this.document.layerPixelsVersion.update((v) => v + 1);
-    this.document.setCanvasSaved(false);
-    this.warpOriginalBuffer = null;
-    this.warpOriginalRect = null;
-    this.warpLayerId = null;
-    this.warpFullLayerBackup = null;
-  }
-
-  private cancelWarp(): void {
-    if (this.warpFullLayerBackup && this.warpLayerId) {
-      const layerBuffer = this.document.getLayerBuffer(this.warpLayerId);
-      if (layerBuffer) {
-        for (let i = 0; i < layerBuffer.length; i++) {
-          layerBuffer[i] = this.warpFullLayerBackup[i];
-        }
-        this.document.layerPixelsVersion.update((v) => v + 1);
-      }
-    }
-
-    this.warp.cancelWarp();
-    this.warpOriginalBuffer = null;
-    this.warpOriginalRect = null;
-    this.warpLayerId = null;
-    this.warpFullLayerBackup = null;
-  }
-
-  private commitPuppetWarp(): void {
-    const state = this.puppetWarp.commitPuppetWarp();
-    if (!state) return;
-
-    if (
-      !this.puppetWarpOriginalBuffer ||
-      !this.puppetWarpOriginalRect ||
-      !this.puppetWarpLayerId ||
-      !this.puppetWarpFullLayerBackup
-    ) {
-      return;
-    }
-
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    const originalLayerBuffer = this.document.getLayerBuffer(
-      this.puppetWarpLayerId,
-    );
-    if (!originalLayerBuffer) return;
-
-    for (let i = 0; i < originalLayerBuffer.length; i++) {
-      originalLayerBuffer[i] = this.puppetWarpFullLayerBackup[i];
-    }
-
-    this.document.layerPixelsVersion.update((v) => v + 1);
-    this.document.setCanvasSaved(false);
-    this.puppetWarpOriginalBuffer = null;
-    this.puppetWarpOriginalRect = null;
-    this.puppetWarpLayerId = null;
-    this.puppetWarpFullLayerBackup = null;
-  }
-
-  private cancelPuppetWarp(): void {
-    if (this.puppetWarpFullLayerBackup && this.puppetWarpLayerId) {
-      const layerBuffer = this.document.getLayerBuffer(this.puppetWarpLayerId);
-      if (layerBuffer) {
-        for (let i = 0; i < layerBuffer.length; i++) {
-          layerBuffer[i] = this.puppetWarpFullLayerBackup[i];
-        }
-        this.document.layerPixelsVersion.update((v) => v + 1);
-      }
-    }
-
-    this.puppetWarp.cancelPuppetWarp();
-    this.puppetWarpOriginalBuffer = null;
-    this.puppetWarpOriginalRect = null;
-    this.puppetWarpLayerId = null;
-    this.puppetWarpFullLayerBackup = null;
-  }
-
-  private renderLiveWarpPreview(): void {
-    const state = this.warp.warpState();
-    if (
-      !state ||
-      !this.warpOriginalBuffer ||
-      !this.warpOriginalRect ||
-      !this.warpLayerId ||
-      !this.warpFullLayerBackup
-    ) {
-      return;
-    }
-
-    const layerBuffer = this.document.getLayerBuffer(this.warpLayerId);
-    if (!layerBuffer) return;
-
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    for (let i = 0; i < layerBuffer.length; i++) {
-      layerBuffer[i] = this.warpFullLayerBackup[i];
-    }
-
-    for (let y = 0; y < this.warpOriginalRect.height; y++) {
-      for (let x = 0; x < this.warpOriginalRect.width; x++) {
-        const destX: number = this.warpOriginalRect.x + x;
-        const destY: number = this.warpOriginalRect.y + y;
-        if (destX >= 0 && destX < canvasW && destY >= 0 && destY < canvasH) {
-          const destIdx = destY * canvasW + destX;
-          layerBuffer[destIdx] = '';
-        }
-      }
-    }
-
-    const dims = this.warp.getGridDimensions();
-    if (!dims) return;
-
-    const result = this.transformService.applyWarp(
-      this.warpOriginalBuffer,
-      this.warpOriginalRect.width,
-      this.warpOriginalRect.height,
-      state.nodes,
-      dims.rows,
-      dims.cols,
-    );
-
-    for (let y = 0; y < result.height; y++) {
-      for (let x = 0; x < result.width; x++) {
-        const srcIdx = y * result.width + x;
-        const color = result.buffer[srcIdx];
-        if (color) {
-          const destX = Math.floor(result.minX + x);
-          const destY = Math.floor(result.minY + y);
-          if (destX >= 0 && destX < canvasW && destY >= 0 && destY < canvasH) {
-            const destIdx = destY * canvasW + destX;
-            layerBuffer[destIdx] = color;
-          }
-        }
-      }
-    }
-
-    this.document.layerPixelsVersion.update((v) => v + 1);
-  }
-
-  private renderLivePuppetWarpPreview(): void {
-    const state = this.puppetWarp.puppetWarpState();
-    if (
-      !state ||
-      !this.puppetWarpOriginalBuffer ||
-      !this.puppetWarpOriginalRect ||
-      !this.puppetWarpLayerId ||
-      !this.puppetWarpFullLayerBackup
-    ) {
-      return;
-    }
-
-    const layerBuffer = this.document.getLayerBuffer(this.puppetWarpLayerId);
-    if (!layerBuffer) return;
-
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    for (let i = 0; i < layerBuffer.length; i++) {
-      layerBuffer[i] = this.puppetWarpFullLayerBackup[i];
-    }
-
-    for (let y = 0; y < this.puppetWarpOriginalRect.height; y++) {
-      for (let x = 0; x < this.puppetWarpOriginalRect.width; x++) {
-        const destX: number = this.puppetWarpOriginalRect.x + x;
-        const destY: number = this.puppetWarpOriginalRect.y + y;
-        if (destX >= 0 && destX < canvasW && destY >= 0 && destY < canvasH) {
-          const destIdx = destY * canvasW + destX;
-          layerBuffer[destIdx] = '';
-        }
-      }
-    }
-
-    if (state.pins.length === 0) {
-      for (let y = 0; y < this.puppetWarpOriginalRect.height; y++) {
-        for (let x = 0; x < this.puppetWarpOriginalRect.width; x++) {
-          const srcIdx = y * this.puppetWarpOriginalRect.width + x;
-          const color = this.puppetWarpOriginalBuffer[srcIdx];
-          if (color) {
-            const destX: number = this.puppetWarpOriginalRect.x + x;
-            const destY: number = this.puppetWarpOriginalRect.y + y;
-            if (
-              destX >= 0 &&
-              destX < canvasW &&
-              destY >= 0 &&
-              destY < canvasH
-            ) {
-              const destIdx = destY * canvasW + destX;
-              layerBuffer[destIdx] = color;
-            }
-          }
-        }
-      }
-      this.document.layerPixelsVersion.update((v) => v + 1);
-      return;
-    }
-
-    const result = this.transformService.applyPuppetWarp(
-      this.puppetWarpOriginalBuffer,
-      this.puppetWarpOriginalRect.width,
-      this.puppetWarpOriginalRect.height,
-      state.pins,
-    );
-
-    for (let y = 0; y < result.height; y++) {
-      for (let x = 0; x < result.width; x++) {
-        const srcIdx = y * result.width + x;
-        const color = result.buffer[srcIdx];
-        if (color) {
-          const destX = Math.floor(result.minX + x);
-          const destY = Math.floor(result.minY + y);
-          if (destX >= 0 && destX < canvasW && destY >= 0 && destY < canvasH) {
-            const destIdx = destY * canvasW + destX;
-            layerBuffer[destIdx] = color;
-          }
-        }
-      }
-    }
-
-    this.document.layerPixelsVersion.update((v) => v + 1);
-  }
-
-  private renderLivePerspectivePreview(): void {
-    const state = this.perspective.perspectiveState();
-    if (
-      !state ||
-      !this.perspectiveOriginalBuffer ||
-      !this.perspectiveOriginalRect ||
-      !this.perspectiveLayerId ||
-      !this.perspectiveFullLayerBackup
-    ) {
-      return;
-    }
-
-    const layerBuffer = this.document.getLayerBuffer(this.perspectiveLayerId);
-    if (!layerBuffer) return;
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    for (let i = 0; i < layerBuffer.length; i++) {
-      layerBuffer[i] = this.perspectiveFullLayerBackup[i];
-    }
-
-    for (let y = 0; y < this.perspectiveOriginalRect.height; y++) {
-      for (let x = 0; x < this.perspectiveOriginalRect.width; x++) {
-        const destX: number = this.perspectiveOriginalRect.x + x;
-        const destY: number = this.perspectiveOriginalRect.y + y;
-        if (destX >= 0 && destX < canvasW && destY >= 0 && destY < canvasH) {
-          const destIdx = destY * canvasW + destX;
-          layerBuffer[destIdx] = '';
-        }
-      }
-    }
-
-    const srcCorners = [
-      { x: 0, y: 0 },
-      { x: state.sourceWidth, y: 0 },
-      { x: state.sourceWidth, y: state.sourceHeight },
-      { x: 0, y: state.sourceHeight },
-    ];
-
-    const dstCorners = [
-      state.corners.topLeft,
-      state.corners.topRight,
-      state.corners.bottomRight,
-      state.corners.bottomLeft,
-    ];
-
-    const result = this.transformService.applyDistort(
-      this.perspectiveOriginalBuffer,
-      this.perspectiveOriginalRect.width,
-      this.perspectiveOriginalRect.height,
-      srcCorners,
-      dstCorners,
-    );
-
-    for (let y = 0; y < result.height; y++) {
-      for (let x = 0; x < result.width; x++) {
-        const srcIdx = y * result.width + x;
-        const color = result.buffer[srcIdx];
-        if (color) {
-          const destX = Math.floor(result.minX + x);
-          const destY = Math.floor(result.minY + y);
-          if (destX >= 0 && destX < canvasW && destY >= 0 && destY < canvasH) {
-            const destIdx = destY * canvasW + destX;
-            layerBuffer[destIdx] = color;
-          }
-        }
-      }
-    }
-
-    this.document.layerPixelsVersion.update((v) => v + 1);
-  }
-
-  private applyTransformToSelection(
-    state: any,
-    scaleX: number,
-    scaleY: number,
-    originalBuffer: string[],
-    originalRect: { x: number; y: number; width: number; height: number },
-  ): void {
-    const layerId = this.freeTransformLayerId;
-    if (!layerId) return;
-    const layerBuffer = this.document.getLayerBuffer(layerId);
-    if (!layerBuffer) return;
-    const canvasWidth = this.document.canvasWidth();
-    const canvasHeight = this.document.canvasHeight();
-
-    const cos = Math.cos((state.rotation * Math.PI) / 180);
-    const sin = Math.sin((state.rotation * Math.PI) / 180);
-    const centerX = originalRect.width / 2;
-    const centerY = originalRect.height / 2;
-
-    const targetW = Math.max(1, Math.round(state.width));
-    const targetH = Math.max(1, Math.round(state.height));
-    const targetX = Math.floor(state.x);
-    const targetY = Math.floor(state.y);
-
-    const transformedBuffer = new Array<string>(targetW * targetH).fill('');
-    for (let y = 0; y < targetH; y++) {
-      for (let x = 0; x < targetW; x++) {
-        const srcX = x / scaleX;
-        const srcY = y / scaleY;
-        const dx = srcX - centerX;
-        const dy = srcY - centerY;
-        const rotX = centerX + dx * cos + dy * sin;
-        const rotY = centerY - dx * sin + dy * cos;
-        let sx = Math.floor(rotX);
-        let sy = Math.floor(rotY);
-        if (this.freeTransformMirrorX) sx = originalRect.width - 1 - sx;
-        if (this.freeTransformMirrorY) sy = originalRect.height - 1 - sy;
-        if (
-          sx >= 0 &&
-          sx < originalRect.width &&
-          sy >= 0 &&
-          sy < originalRect.height
-        ) {
-          const srcIdx = sy * originalRect.width + sx;
-          const destIdx = y * targetW + x;
-          transformedBuffer[destIdx] = originalBuffer[srcIdx] || '';
-        }
-      }
-    }
-    for (let y = 0; y < targetH; y++) {
-      for (let x = 0; x < targetW; x++) {
-        const destX = targetX + x;
-        const destY = targetY + y;
-        if (
-          destX >= 0 &&
-          destX < canvasWidth &&
-          destY >= 0 &&
-          destY < canvasHeight
-        ) {
-          const srcIdx = y * targetW + x;
-          const destIdx = destY * canvasWidth + destX;
-          const col = transformedBuffer[srcIdx];
-          if (col) layerBuffer[destIdx] = col;
-        }
-      }
-    }
-    this.document.layerPixelsVersion.update((v) => v + 1);
-  }
-
-  private restoreOriginalTransformPixels(): void {
-    if (
-      !this.freeTransformOriginalBuffer ||
-      !this.freeTransformOriginalRect ||
-      !this.freeTransformLayerId
-    )
-      return;
-    const layerBuffer = this.document.getLayerBuffer(this.freeTransformLayerId);
-    if (!layerBuffer) return;
-    const canvasWidth = this.document.canvasWidth();
-    const canvasHeight = this.document.canvasHeight();
-    const rect = this.freeTransformOriginalRect;
-    for (let y = 0; y < rect.height; y++) {
-      for (let x = 0; x < rect.width; x++) {
-        const srcIdx = y * rect.width + x;
-        const destX = rect.x + x;
-        const destY = rect.y + y;
-        if (
-          destX >= 0 &&
-          destX < canvasWidth &&
-          destY >= 0 &&
-          destY < canvasHeight
-        ) {
-          const destIdx = destY * canvasWidth + destX;
-          const col = this.freeTransformOriginalBuffer[srcIdx];
-          if (col) layerBuffer[destIdx] = col;
-        }
-      }
-    }
-    this.document.layerPixelsVersion.update((v) => v + 1);
-  }
-
-  // Live preview: apply transformed pixels directly to layer buffer while dragging,
-  // restoring previous preview first to avoid accumulation.
-  private renderLiveFreeTransformPreview(): void {
-    const state = this.freeTransform.transformState();
-    if (
-      !state ||
-      !this.freeTransformOriginalBuffer ||
-      !this.freeTransformOriginalRect ||
-      !this.freeTransformLayerId
-    )
-      return;
-    const layerBuffer = this.document.getLayerBuffer(this.freeTransformLayerId);
-    if (!layerBuffer) return;
-    const canvasW = this.document.canvasWidth();
-    const canvasH = this.document.canvasHeight();
-
-    // Restore pixels from backup before clearing preview.
-    // This ensures pixels outside the original selection are preserved.
-    if (
-      this.freeTransformPreviewIndices &&
-      this.freeTransformPreviewIndices.length &&
-      this.freeTransformPreviewBackup
-    ) {
-      for (const idx of this.freeTransformPreviewIndices) {
-        if (idx >= 0 && idx < layerBuffer.length) {
-          const backupValue = this.freeTransformPreviewBackup.get(idx);
-          layerBuffer[idx] = backupValue !== undefined ? backupValue : '';
-        }
-      }
-    }
-    this.freeTransformPreviewIndices = [];
-    this.freeTransformPreviewBackup = new Map<number, string>();
-
-    const srcW = this.freeTransformOriginalRect.width;
-    const srcH = this.freeTransformOriginalRect.height;
-    const scaleX = state.width / srcW;
-    const scaleY = state.height / srcH;
-    const rad = (state.rotation * Math.PI) / 180;
-    const cos = Math.cos(rad);
-    const sin = Math.sin(rad);
-    const cX = srcW / 2;
-    const cY = srcH / 2;
-    const tgtW = Math.max(1, Math.round(state.width));
-    const tgtH = Math.max(1, Math.round(state.height));
-    const baseX = Math.floor(state.x);
-    const baseY = Math.floor(state.y);
-
-    for (let y = 0; y < tgtH; y++) {
-      for (let x = 0; x < tgtW; x++) {
-        const srcX = x / scaleX;
-        const srcY = y / scaleY;
-        const dx = srcX - cX;
-        const dy = srcY - cY;
-        const rotX = cX + dx * cos + dy * sin;
-        const rotY = cY - dx * sin + dy * cos;
-        let sx = Math.floor(rotX);
-        let sy = Math.floor(rotY);
-        if (this.freeTransformMirrorX) sx = srcW - 1 - sx;
-        if (this.freeTransformMirrorY) sy = srcH - 1 - sy;
-        if (sx >= 0 && sx < srcW && sy >= 0 && sy < srcH) {
-          const srcIdx = sy * srcW + sx;
-          const col = this.freeTransformOriginalBuffer[srcIdx];
-          if (col && col.length) {
-            const destX = baseX + x;
-            const destY = baseY + y;
-            if (
-              destX >= 0 &&
-              destX < canvasW &&
-              destY >= 0 &&
-              destY < canvasH
-            ) {
-              const destIdx = destY * canvasW + destX;
-              // Backup the current pixel value before overwriting
-              if (!this.freeTransformPreviewBackup.has(destIdx)) {
-                this.freeTransformPreviewBackup.set(
-                  destIdx,
-                  layerBuffer[destIdx] || '',
-                );
-              }
-              layerBuffer[destIdx] = col;
-              this.freeTransformPreviewIndices.push(destIdx);
-            }
-          }
-        }
-      }
-    }
-    this.document.layerPixelsVersion.update((v) => v + 1);
-  }
-
-  private isPixelInSelection(x: number, y: number): boolean {
-    const sel = this.document.selectionRect();
-    const shape = this.document.selectionShape();
-    const poly = this.document.selectionPolygon();
-    return this.document.isPixelWithinSelection(x, y, sel, shape, poly);
-  }
+  resetRotation() { this.rotation.set(0); }
+  increaseZoom(step = 0.1) { const factor = 1 + Math.max(0, step); this.applyZoom(this.scale() * factor); }
+  decreaseZoom(step = 0.1) { const factor = 1 + Math.max(0, step); this.applyZoom(this.scale() / factor); }
 
   ngOnDestroy(): void {
     this.viewReady.set(false);
     this.closeContextMenu();
-    if (this.resizeObserver) {
-      try {
-        this.resizeObserver.disconnect();
-      } catch {}
-      this.resizeObserver = null;
-    }
+    if (this.resizeObserver) { try { this.resizeObserver.disconnect(); } catch {} this.resizeObserver = null; }
     if (this.stopRenderEffect) {
       try {
-        if ((this.stopRenderEffect as any).destroy) {
-          (this.stopRenderEffect as any).destroy();
-        } else if (typeof (this.stopRenderEffect as any) === 'function') {
-          (this.stopRenderEffect as any)();
-        }
+        if ((this.stopRenderEffect as any).destroy) (this.stopRenderEffect as any).destroy();
+        else if (typeof (this.stopRenderEffect as any) === 'function') (this.stopRenderEffect as any)();
       } catch {}
       this.stopRenderEffect = null;
     }
-
     if (this.resizeListener && typeof window !== 'undefined') {
-      try {
-        window.removeEventListener(
-          'resize',
-          this.resizeListener as EventListener,
-        );
-      } catch {}
+      try { window.removeEventListener('resize', this.resizeListener as EventListener); } catch {}
       this.resizeListener = null;
     }
     if (this.keyListener && typeof window !== 'undefined') {
-      try {
-        window.removeEventListener(
-          'keydown',
-          this.keyListener as EventListener,
-        );
-      } catch {}
+      try { window.removeEventListener('keydown', this.keyListener as EventListener); } catch {}
       this.keyListener = null;
     }
     this.generationService.clearGenerationInterval();
@@ -3237,18 +816,11 @@ export class EditorCanvas implements OnDestroy {
       if (!canvas) return;
       const w = Math.max(1, this.document.canvasWidth());
       const h = Math.max(1, this.document.canvasHeight());
-
-      const { contentWidth, contentHeight, paddingLeft, paddingTop } =
-        this.measureContainer();
+      const { contentWidth, contentHeight, paddingLeft, paddingTop } = this.measureContainer();
       if (contentWidth <= 0 || contentHeight <= 0) return;
-
-      const fitScale = Math.max(
-        this.minScale,
-        Math.min(contentWidth / w, contentHeight / h),
-      );
+      const fitScale = Math.max(this.minScale, Math.min(contentWidth / w, contentHeight / h));
       const initialScale = Math.min(this.maxScale, fitScale);
       this.scale.set(initialScale);
-
       const displayWidth = w * initialScale;
       const displayHeight = h * initialScale;
       const offsetX = paddingLeft + (contentWidth - displayWidth) / 2;
@@ -3256,1533 +828,94 @@ export class EditorCanvas implements OnDestroy {
       this.panX.set(offsetX);
       this.panY.set(offsetY);
       this.updateTileSize(this.tools.brushSize());
-    } catch (e) {
-      // best-effort: ignore errors
-    }
+    } catch (e) {}
   }
 
-  private applyZoom(
-    nextScale: number,
-    anchor?: { clientX: number; clientY: number },
-  ) {
+  private applyZoom(nextScale: number, anchor?: { clientX: number; clientY: number }) {
     const clamped = Math.min(this.maxScale, Math.max(this.minScale, nextScale));
     const prev = this.scale();
-    if (!this.canvasEl?.nativeElement) {
-      this.scale.set(clamped);
-      this.updateTileSize(this.tools.brushSize());
-      return;
-    }
-    if (Math.abs(clamped - prev) < 0.0001) {
-      return;
-    }
-
+    if (!this.canvasEl?.nativeElement) { this.scale.set(clamped); this.updateTileSize(this.tools.brushSize()); return; }
+    if (Math.abs(clamped - prev) < 0.0001) return;
     const container = this.canvasContainer?.nativeElement;
     const containerRect = container ? container.getBoundingClientRect() : null;
     const prevPanX = this.panX();
     const prevPanY = this.panY();
-
-    const pivotX =
-      anchor?.clientX ??
-      (containerRect ? containerRect.left + containerRect.width / 2 : 0);
-    const pivotY =
-      anchor?.clientY ??
-      (containerRect ? containerRect.top + containerRect.height / 2 : 0);
+    const pivotX = anchor?.clientX ?? (containerRect ? containerRect.left + containerRect.width / 2 : 0);
+    const pivotY = anchor?.clientY ?? (containerRect ? containerRect.top + containerRect.height / 2 : 0);
     const containerOffsetX = containerRect ? pivotX - containerRect.left : 0;
     const containerOffsetY = containerRect ? pivotY - containerRect.top : 0;
     const worldX = containerRect ? (containerOffsetX - prevPanX) / prev : 0;
     const worldY = containerRect ? (containerOffsetY - prevPanY) / prev : 0;
-
     this.scale.set(clamped);
-
     if (containerRect) {
       const newOffsetX = worldX * clamped;
       const newOffsetY = worldY * clamped;
       this.panX.set(containerOffsetX - newOffsetX);
       this.panY.set(containerOffsetY - newOffsetY);
     }
-
     this.updateTileSize(this.tools.brushSize());
   }
 
   private drawCanvas() {
     const canvas = this.canvasEl?.nativeElement;
     if (!canvas) return;
-    const w = this.document.canvasWidth();
-    const h = this.document.canvasHeight();
-    const scale = this.scale();
-    const currentPanX = this.panX();
-    const currentPanY = this.panY();
-    const dpr =
-      typeof window !== 'undefined' ? window.devicePixelRatio || 1 : 1;
-
-    this.workspaceWidth();
-    this.workspaceHeight();
-    const container = this.canvasContainer?.nativeElement;
-    const containerWidth = container ? container.clientWidth : w * scale;
-    const containerHeight = container ? container.clientHeight : h * scale;
-
-    const pixelWidth = Math.max(1, Math.floor(containerWidth * dpr));
-    const pixelHeight = Math.max(1, Math.floor(containerHeight * dpr));
-    if (canvas.width !== pixelWidth) canvas.width = pixelWidth;
-    if (canvas.height !== pixelHeight) canvas.height = pixelHeight;
-
-    const ctx = canvas.getContext('2d');
-    if (!ctx) return;
-    const root =
-      typeof document !== 'undefined' ? document.documentElement : null;
-    const isDark = !!root && root.classList.contains('dark');
-    ctx.setTransform(1, 0, 0, 1, 0, 0);
-    ctx.clearRect(0, 0, pixelWidth, pixelHeight);
-
-    ctx.fillStyle = isDark ? '#1f2937' : '#e5e7eb';
-    ctx.fillRect(0, 0, pixelWidth, pixelHeight);
-
-    ctx.setTransform(scale * dpr, 0, 0, scale * dpr, currentPanX * dpr, currentPanY * dpr);
-    ctx.imageSmoothingEnabled = false;
-    const pxLineWidth = 1 / (scale * dpr);
-
-    const tile = this.tileSize();
-    const darkTile = isDark ? 'rgba(0,0,0,0.06)' : 'rgba(0,0,0,0.04)';
-    const lightTile = isDark
-      ? 'rgba(255,255,255,0.03)'
-      : 'rgba(255,255,255,0.02)';
-    ctx.save();
-    ctx.fillStyle = isDark ? '#374151' : '#ffffff';
-    ctx.fillRect(0, 0, w, h);
-    for (let y = 0; y < h; y += tile) {
-      for (let x = 0; x < w; x += tile) {
-        const even = ((x / tile + y / tile) & 1) === 0;
-        ctx.fillStyle = even ? lightTile : darkTile;
-        ctx.fillRect(x, y, Math.min(tile, w - x), Math.min(tile, h - y));
-      }
-    }
-    ctx.restore();
-
-    this.document.layerPixelsVersion();
-
-    const frameId = this.getCurrentFrameId();
-    const currentAnim = this.document.getCurrentAnimation();
-    const animationId = currentAnim?.id;
-    const currentTime = this.document.keyframeService?.getCurrentTime();
-    const isInTimelineMode = this.document.keyframeService?.isTimelineMode();
-    const shouldApplyBoneTransforms = !!(
-      animationId &&
-      typeof currentTime === 'number' &&
-      isInTimelineMode
-    );
-
-    const layers = this.document.getFlattenedLayers();
-    for (let li = layers.length - 1; li >= 0; li--) {
-      const layer = layers[li];
-      if (!layer.visible) continue;
-      const buf = this.document.getLayerBuffer(layer.id);
-      if (!buf || buf.length !== w * h) continue;
-      ctx.save();
-
-      if (shouldApplyBoneTransforms) {
-        const boundSourcePixels = new Set<number>();
-        const destinationPixelMap = new Map<
-          number,
-          { color: string; priority: number }
-        >();
-        const bindings =
-          this.document.keyframeService.getPixelBindings(frameId);
-
-        const transformCache = new Map<string, any>();
-        const pixelBindingsMap = new Map<number, typeof bindings>();
-
-        for (const binding of bindings) {
-          const sourceIdx = binding.pixelY * w + binding.pixelX;
-          const existingBindings = pixelBindingsMap.get(sourceIdx) || [];
-          existingBindings.push(binding);
-          pixelBindingsMap.set(sourceIdx, existingBindings);
-        }
-
-        for (const [sourceIdx, pixelBindings] of pixelBindingsMap.entries()) {
-          const col = buf[sourceIdx];
-          if (!col || !col.length) continue;
-
-          const dominantBinding = pixelBindings.reduce((prev, current) =>
-            current.weight > prev.weight ? current : prev,
-          );
-
-          const cacheKey = `${dominantBinding.boneId}:${dominantBinding.bonePointId}`;
-          let transform = transformCache.get(cacheKey);
-
-          if (!transform) {
-            transform = this.document.keyframeService.interpolateBoneTransform(
-              animationId,
-              dominantBinding.boneId,
-              dominantBinding.bonePointId,
-              currentTime,
-            );
-            transformCache.set(cacheKey, transform);
-          }
-
-          if (transform) {
-            const transformedX = Math.round(
-              transform.x + dominantBinding.offsetX,
-            );
-            const transformedY = Math.round(
-              transform.y + dominantBinding.offsetY,
-            );
-
-            if (
-              transformedX >= 0 &&
-              transformedX < w &&
-              transformedY >= 0 &&
-              transformedY < h
-            ) {
-              const destIdx = transformedY * w + transformedX;
-              const distSq =
-                dominantBinding.offsetX * dominantBinding.offsetX +
-                dominantBinding.offsetY * dominantBinding.offsetY;
-              const priority = -distSq;
-
-              const existing = destinationPixelMap.get(destIdx);
-              if (!existing || priority > existing.priority) {
-                destinationPixelMap.set(destIdx, { color: col, priority });
-              }
-              boundSourcePixels.add(sourceIdx);
-            }
-          } else {
-            const pixelY = Math.floor(sourceIdx / w);
-            const pixelX = sourceIdx % w;
-            const destIdx = pixelY * w + pixelX;
-            destinationPixelMap.set(destIdx, { color: col, priority: 0 });
-            boundSourcePixels.add(sourceIdx);
-          }
-        }
-
-        const bones = this.boneService.getBones(frameId);
-        for (const bone of bones) {
-          for (const point of bone.points) {
-            if (!point.parentId) continue;
-
-            const parent = bone.points.find((p) => p.id === point.parentId);
-            if (!parent) continue;
-
-            const pointTransform =
-              this.document.keyframeService.interpolateBoneTransform(
-                animationId,
-                bone.id,
-                point.id,
-                currentTime,
-              );
-            const parentTransform =
-              this.document.keyframeService.interpolateBoneTransform(
-                animationId,
-                bone.id,
-                parent.id,
-                currentTime,
-              );
-
-            const p1x = pointTransform ? pointTransform.x : point.x;
-            const p1y = pointTransform ? pointTransform.y : point.y;
-            const p2x = parentTransform ? parentTransform.x : parent.x;
-            const p2y = parentTransform ? parentTransform.y : parent.y;
-
-            const dx = p2x - p1x;
-            const dy = p2y - p1y;
-            const dist = Math.sqrt(dx * dx + dy * dy);
-            const steps = Math.ceil(dist);
-
-            for (let i = 0; i <= steps; i++) {
-              const t = steps > 0 ? i / steps : 0;
-              const interpX = Math.round(p1x + dx * t);
-              const interpY = Math.round(p1y + dy * t);
-
-              if (interpX >= 0 && interpX < w && interpY >= 0 && interpY < h) {
-                const interpIdx = interpY * w + interpX;
-                const sourcePixelX = Math.round(
-                  point.x + (parent.x - point.x) * t,
-                );
-                const sourcePixelY = Math.round(
-                  point.y + (parent.y - point.y) * t,
-                );
-
-                if (
-                  sourcePixelX >= 0 &&
-                  sourcePixelX < w &&
-                  sourcePixelY >= 0 &&
-                  sourcePixelY < h
-                ) {
-                  const sourceIdx = sourcePixelY * w + sourcePixelX;
-                  const sourceColor = buf[sourceIdx];
-
-                  if (
-                    sourceColor &&
-                    sourceColor.length &&
-                    !destinationPixelMap.has(interpIdx)
-                  ) {
-                    destinationPixelMap.set(interpIdx, {
-                      color: sourceColor,
-                      priority: -1000,
-                    });
-                  }
-                }
-              }
-            }
-          }
-        }
-
-        for (let yy = 0; yy < h; yy++) {
-          for (let xx = 0; xx < w; xx++) {
-            const idx = yy * w + xx;
-
-            if (destinationPixelMap.has(idx)) {
-              const pixelData = destinationPixelMap.get(idx);
-              if (pixelData) {
-                ctx.fillStyle = pixelData.color;
-                ctx.fillRect(xx, yy, 1, 1);
-              }
-            } else if (!boundSourcePixels.has(idx)) {
-              const col = buf[idx];
-              if (col && col.length) {
-                ctx.fillStyle = col;
-                ctx.fillRect(xx, yy, 1, 1);
-              }
-            }
-          }
-        }
-      } else {
-        for (let yy = 0; yy < h; yy++) {
-          for (let xx = 0; xx < w; xx++) {
-            const col = buf[yy * w + xx];
-            if (col && col.length) {
-              ctx.fillStyle = col;
-              ctx.fillRect(xx, yy, 1, 1);
-            }
-          }
-        }
-      }
-      ctx.restore();
-    }
-
-    if (this.movingContentBuffer && this.movingContentOriginalRect) {
-      const sel = this.document.selectionRect();
-      if (sel) {
-        const dx = sel.x - this.movingContentOriginalRect.x;
-        const dy = sel.y - this.movingContentOriginalRect.y;
-        ctx.save();
-        for (let y = 0; y < h; y++) {
-          for (let x = 0; x < w; x++) {
-            const idx = y * w + x;
-            const col = this.movingContentBuffer[idx];
-            if (col && col.length) {
-              const newX = x + dx;
-              const newY = y + dy;
-              if (newX >= 0 && newX < w && newY >= 0 && newY < h) {
-                ctx.fillStyle = col;
-                ctx.fillRect(newX, newY, 1, 1);
-              }
-            }
-          }
-        }
-        ctx.restore();
-      }
-    }
-
-    const activeShape = this.activeShapeTool();
-    const shapeStart = this.shapeStart();
-    const shapeCurrent = this.shapeCurrent();
-    if (activeShape && shapeStart && shapeCurrent) {
-      ctx.save();
-      if (activeShape === 'line') {
-        ctx.strokeStyle = this.tools.lineColor();
-        ctx.lineWidth = Math.max(pxLineWidth, this.tools.lineThickness());
-        ctx.lineCap = 'round';
-        ctx.beginPath();
-        ctx.moveTo(shapeStart.x + 0.5, shapeStart.y + 0.5);
-        ctx.lineTo(shapeCurrent.x + 0.5, shapeCurrent.y + 0.5);
-        ctx.stroke();
-      } else {
-        if (activeShape === 'square') {
-          const bounds = this.computeRectBounds(
-            shapeStart,
-            shapeCurrent,
-            this.shapeConstrainUniform(),
-          );
-          this.renderSquarePreview(
-            ctx,
-            bounds,
-            this.getSquareDrawOptions(),
-            pxLineWidth,
-          );
-        } else {
-          const bounds = this.computeRectBounds(
-            shapeStart,
-            shapeCurrent,
-            this.shapeConstrainUniform(),
-          );
-          this.renderEllipsePreview(
-            ctx,
-            bounds,
-            this.getCircleDrawOptions(),
-            pxLineWidth,
-          );
-        }
-      }
-      ctx.restore();
-    }
-
-    const penDrawing = this.penDrawing();
-    const penPoints = this.penPoints();
-    if (penDrawing && penPoints.length > 0) {
-      ctx.save();
-      ctx.strokeStyle = this.tools.penColor();
-      ctx.lineWidth = Math.max(pxLineWidth, this.tools.penThickness());
-      ctx.lineCap = 'round';
-      ctx.lineJoin = 'round';
-      const lineMode = this.tools.penLineMode();
-      if (lineMode === 'spline' && penPoints.length >= 3) {
-        const splinePoints = this.shapeService.catmullRomSpline(penPoints, 10);
-        if (splinePoints.length > 1) {
-          ctx.beginPath();
-          ctx.moveTo(splinePoints[0].x + 0.5, splinePoints[0].y + 0.5);
-          for (let i = 1; i < splinePoints.length; i++) {
-            ctx.lineTo(splinePoints[i].x + 0.5, splinePoints[i].y + 0.5);
-          }
-          ctx.stroke();
-        }
-      } else {
-        ctx.beginPath();
-        ctx.moveTo(penPoints[0].x + 0.5, penPoints[0].y + 0.5);
-        for (let i = 1; i < penPoints.length; i++) {
-          ctx.lineTo(penPoints[i].x + 0.5, penPoints[i].y + 0.5);
-        }
-        ctx.stroke();
-      }
-      for (const point of penPoints) {
-        ctx.fillStyle = this.tools.penColor();
-        ctx.beginPath();
-        ctx.arc(point.x + 0.5, point.y + 0.5, Math.max(2, this.tools.penThickness() / 2), 0, Math.PI * 2);
-        ctx.fill();
-      }
-      ctx.restore();
-    }
-
-    const tool = this.tools.currentTool();
-    if (tool === 'brush' || tool === 'eraser') {
-      const mouseXVal = this.mouseX();
-      const mouseYVal = this.mouseY();
-      if (mouseXVal !== null && mouseYVal !== null) {
-        const logicalMouseX = (mouseXVal - currentPanX) / scale;
-        const logicalMouseY = (mouseYVal - currentPanY) / scale;
-        const size =
-          tool === 'eraser' ? this.tools.eraserSize() : this.tools.brushSize();
-        const bSize = Math.max(1, size);
-        const half = Math.floor((bSize - 1) / 2);
-        const brushX = Math.floor(logicalMouseX) - half;
-        const brushY = Math.floor(logicalMouseY) - half;
-
-        ctx.save();
-        ctx.lineWidth = pxLineWidth;
-        if (tool === 'eraser') {
-          ctx.fillStyle = isDark
-            ? 'rgba(255,255,255,0.12)'
-            : 'rgba(0,0,0,0.10)';
-          ctx.fillRect(brushX, brushY, bSize, bSize);
-          ctx.strokeStyle = isDark
-            ? 'rgba(255,255,255,0.5)'
-            : 'rgba(0,0,0,0.5)';
-          ctx.strokeRect(
-            brushX + 0.5,
-            brushY + 0.5,
-            Math.max(0, bSize - 1),
-            Math.max(0, bSize - 1),
-          );
-        } else {
-          ctx.fillStyle = this.tools.brushColor();
-          ctx.globalAlpha = 0.9;
-          ctx.fillRect(brushX, brushY, bSize, bSize);
-          ctx.globalAlpha = 1;
-          ctx.strokeStyle = isDark
-            ? 'rgba(255,255,255,0.6)'
-            : 'rgba(0,0,0,0.6)';
-          ctx.strokeRect(
-            brushX + 0.5,
-            brushY + 0.5,
-            Math.max(0, bSize - 1),
-            Math.max(0, bSize - 1),
-          );
-        }
-        ctx.restore();
-      }
-    }
-
-    if (tool === 'bone') {
-      const bones = this.boneService.getBones(frameId);
-
-      const getPointPosition = (
-        boneId: string,
-        point: BonePoint,
-      ): { x: number; y: number } => {
-        if (shouldApplyBoneTransforms) {
-          const transform =
-            this.document.keyframeService.interpolateBoneTransform(
-              animationId,
-              boneId,
-              point.id,
-              currentTime,
-            );
-          if (transform) {
-            return { x: transform.x, y: transform.y };
-          }
-        }
-        return { x: point.x, y: point.y };
-      };
-
-      ctx.save();
-      for (const bone of bones) {
-        if (bone.points.length === 0) continue;
-
-        ctx.strokeStyle = bone.color;
-        ctx.lineWidth = Math.max(pxLineWidth, bone.thickness);
-        ctx.lineCap = 'round';
-        ctx.lineJoin = 'round';
-
-        const drawnConnections = new Set<string>();
-
-        const drawConnection = (point: BonePoint) => {
-          if (point.parentId) {
-            const connectionKey = `${point.parentId}->${point.id}`;
-            if (drawnConnections.has(connectionKey)) return;
-
-            const parent = bone.points.find((p) => p.id === point.parentId);
-            if (parent) {
-              const parentPos = getPointPosition(bone.id, parent);
-              const pointPos = getPointPosition(bone.id, point);
-              ctx.beginPath();
-              ctx.moveTo(parentPos.x + 0.5, parentPos.y + 0.5);
-              ctx.lineTo(pointPos.x + 0.5, pointPos.y + 0.5);
-              ctx.stroke();
-              drawnConnections.add(connectionKey);
-            }
-          }
-        };
-
-        for (const point of bone.points) {
-          drawConnection(point);
-        }
-
-        for (const point of bone.points) {
-          const isSelected = this.boneService.getSelectedPoint() === point.id;
-          const radius = Math.max(3 / scale, 0.5);
-          const pos = getPointPosition(bone.id, point);
-
-          ctx.fillStyle = point.color || bone.color;
-          ctx.beginPath();
-          ctx.arc(pos.x + 0.5, pos.y + 0.5, radius, 0, Math.PI * 2);
-          ctx.fill();
-
-          if (isSelected) {
-            ctx.strokeStyle = '#ffffff';
-            ctx.lineWidth = Math.max(pxLineWidth * 2, 0.3);
-            ctx.stroke();
-          }
-
-          ctx.strokeStyle = isDark
-            ? 'rgba(255,255,255,0.8)'
-            : 'rgba(0,0,0,0.8)';
-          ctx.lineWidth = Math.max(pxLineWidth, 0.2);
-          ctx.stroke();
-        }
-      }
-      ctx.restore();
-
-      const bindings = this.document.keyframeService?.getPixelBindings(frameId);
-      if (bindings && bindings.length > 0 && this.tools.boneAutoBindEnabled()) {
-        ctx.save();
-
-        const getBoneDisplayColor = (boneId: string): string => {
-          let hash = 0;
-          for (let i = 0; i < boneId.length; i++) {
-            hash = (hash << 5) - hash + boneId.charCodeAt(i);
-            hash = hash & hash;
-          }
-
-          const hue = Math.abs(hash) % 360;
-          const saturation = 60 + (Math.abs(hash >> 8) % 20);
-          const lightness = 50 + (Math.abs(hash >> 16) % 15);
-
-          return `hsl(${hue}, ${saturation}%, ${lightness}%)`;
-        };
-
-        const boneColorCache = new Map<
-          string,
-          { r: number; g: number; b: number }
-        >();
-
-        const pixelBindingCounts = new Map<string, number>();
-        for (const binding of bindings) {
-          const key = `${binding.pixelX},${binding.pixelY}`;
-          pixelBindingCounts.set(key, (pixelBindingCounts.get(key) || 0) + 1);
-        }
-
-        for (const binding of bindings) {
-          const pixelKey = `${binding.pixelX},${binding.pixelY}`;
-          const bindingCount = pixelBindingCounts.get(pixelKey) || 1;
-          const isMultiBound = bindingCount > 1;
-
-          let rgb = boneColorCache.get(binding.boneId);
-
-          if (!rgb) {
-            const hslColor = getBoneDisplayColor(binding.boneId);
-            const tempDiv =
-              typeof document !== 'undefined'
-                ? document.createElement('div')
-                : null;
-            if (tempDiv) {
-              tempDiv.style.color = hslColor;
-              document.body.appendChild(tempDiv);
-              const computed = getComputedStyle(tempDiv).color;
-              document.body.removeChild(tempDiv);
-
-              const match = computed.match(/rgb\((\d+),\s*(\d+),\s*(\d+)\)/);
-              if (match) {
-                rgb = {
-                  r: parseInt(match[1]),
-                  g: parseInt(match[2]),
-                  b: parseInt(match[3]),
-                };
-              } else {
-                rgb = { r: 255, g: 102, b: 0 };
-              }
-            } else {
-              const hue =
-                Math.abs(
-                  binding.boneId
-                    .split('')
-                    .reduce((a, c) => a + c.charCodeAt(0), 0),
-                ) % 360;
-              const s = 0.7;
-              const l = 0.6;
-              const c = (1 - Math.abs(2 * l - 1)) * s;
-              const x = c * (1 - Math.abs(((hue / 60) % 2) - 1));
-              const m = l - c / 2;
-              let r = 0,
-                g = 0,
-                b = 0;
-              if (hue < 60) {
-                r = c;
-                g = x;
-              } else if (hue < 120) {
-                r = x;
-                g = c;
-              } else if (hue < 180) {
-                g = c;
-                b = x;
-              } else if (hue < 240) {
-                g = x;
-                b = c;
-              } else if (hue < 300) {
-                r = x;
-                b = c;
-              } else {
-                r = c;
-                b = x;
-              }
-              rgb = {
-                r: Math.round((r + m) * 255),
-                g: Math.round((g + m) * 255),
-                b: Math.round((b + m) * 255),
-              };
-            }
-            if (rgb) {
-              boneColorCache.set(binding.boneId, rgb);
-            }
-          }
-
-          if (rgb) {
-            if (isMultiBound) {
-              ctx.fillStyle = `rgba(255, 255, 0, 0.4)`;
-            } else {
-              ctx.fillStyle = `rgba(${rgb.r}, ${rgb.g}, ${rgb.b}, 0.2)`;
-            }
-            ctx.fillRect(binding.pixelX, binding.pixelY, 1, 1);
-          }
-        }
-
-        ctx.restore();
-      }
-    }
-
-    // Draw active selection if present (but hide when Free Transform, Distort, Perspective, Warp, or Puppet Warp is active)
-    const sel = this.document.selectionRect();
-    const selShape = this.document.selectionShape();
-    const isFreeTransformActive = this.freeTransform.isActive();
-    const isDistortActive = this.distort.isActive();
-    const isPerspectiveActive = this.perspective.isActive();
-    const isWarpActive = this.warp.isActive();
-    const isPuppetWarpActive = this.puppetWarp.isActive();
-    if (
-      sel &&
-      sel.width > 0 &&
-      sel.height > 0 &&
-      !isFreeTransformActive &&
-      !isDistortActive &&
-      !isPerspectiveActive &&
-      !isWarpActive &&
-      !isPuppetWarpActive
-    ) {
-      ctx.save();
-
-      // Check if we have a mask-based selection
-      const mask = this.document.selectionMask();
-
-      if (mask) {
-        // Draw mask-based selection by rendering individual pixels
-        ctx.fillStyle = isDark ? 'rgba(255,255,255,0.04)' : 'rgba(0,0,0,0.04)';
-        for (const key of mask) {
-          const [xStr, yStr] = key.split(',');
-          const x = parseInt(xStr, 10);
-          const y = parseInt(yStr, 10);
-          ctx.fillRect(x, y, 1, 1);
-        }
-
-        // Draw marching ants border by detecting edges
-        // An edge exists where a selected pixel borders an unselected pixel
-        ctx.setLineDash([4 / scale, 3 / scale]);
-        ctx.strokeStyle = isDark ? 'rgba(255,255,255,0.8)' : 'rgba(0,0,0,0.8)';
-        ctx.lineWidth = pxLineWidth;
-        ctx.beginPath();
-
-        for (const key of mask) {
-          const [xStr, yStr] = key.split(',');
-          const x = parseInt(xStr, 10);
-          const y = parseInt(yStr, 10);
-
-          // Check all 4 neighbors to see if we need to draw an edge
-          // Top edge
-          if (!mask.has(`${x},${y - 1}`)) {
-            ctx.moveTo(x, y);
-            ctx.lineTo(x + 1, y);
-          }
-          // Right edge
-          if (!mask.has(`${x + 1},${y}`)) {
-            ctx.moveTo(x + 1, y);
-            ctx.lineTo(x + 1, y + 1);
-          }
-          // Bottom edge
-          if (!mask.has(`${x},${y + 1}`)) {
-            ctx.moveTo(x, y + 1);
-            ctx.lineTo(x + 1, y + 1);
-          }
-          // Left edge
-          if (!mask.has(`${x - 1},${y}`)) {
-            ctx.moveTo(x, y);
-            ctx.lineTo(x, y + 1);
-          }
-        }
-
-        ctx.stroke();
-      } else {
-        // translucent fill
-        ctx.fillStyle = isDark ? 'rgba(255,255,255,0.04)' : 'rgba(0,0,0,0.04)';
-        if (selShape === 'ellipse') {
-          const cx = sel.x + sel.width / 2 - 0.5;
-          const cy = sel.y + sel.height / 2 - 0.5;
-          const rx = Math.max(0.5, sel.width / 2);
-          const ry = Math.max(0.5, sel.height / 2);
-          ctx.beginPath();
-          ctx.ellipse(cx, cy, rx, ry, 0, 0, Math.PI * 2);
-          ctx.fill();
-          // dashed stroke
-          ctx.setLineDash([4 / scale, 3 / scale]);
-          ctx.strokeStyle = isDark
-            ? 'rgba(255,255,255,0.8)'
-            : 'rgba(0,0,0,0.8)';
-          ctx.lineWidth = pxLineWidth;
-          ctx.stroke();
-        } else if (selShape === 'lasso') {
-          const poly = this.document.selectionPolygon();
-          if (poly && poly.length > 0) {
-            ctx.beginPath();
-            ctx.moveTo(poly[0].x + 0.5, poly[0].y + 0.5);
-            for (let i = 1; i < poly.length; i++) {
-              ctx.lineTo(poly[i].x + 0.5, poly[i].y + 0.5);
-            }
-            // Optionally close the path for visual completeness
-            ctx.closePath();
-            ctx.fill();
-            ctx.setLineDash([4 / scale, 3 / scale]);
-            ctx.strokeStyle = isDark
-              ? 'rgba(255,255,255,0.8)'
-              : 'rgba(0,0,0,0.8)';
-            ctx.lineWidth = pxLineWidth;
-            ctx.stroke();
-          }
-        } else {
-          // ctx.fillRect(sel.x, sel.y, sel.width, sel.height);
-          // dashed stroke
-          ctx.setLineDash([4 / scale, 3 / scale]);
-          ctx.strokeStyle = isDark
-            ? 'rgba(255,255,255,0.8)'
-            : 'rgba(0,0,0,0.8)';
-          ctx.lineWidth = pxLineWidth;
-          ctx.strokeRect(
-            sel.x,
-            sel.y,
-            Math.max(0, sel.width),
-            Math.max(0, sel.height),
-          );
-        }
-      }
-      ctx.restore();
-    }
-
-    const freeTransformState = this.freeTransform.transformState();
-    if (freeTransformState) {
-      // removed duplicate overlay drawing; handled in the dedicated overlay section below
-    }
-
-    // Free transform overlay (selection live preview)
-    const ftState = this.freeTransform.transformState();
-    if (ftState && this.document.selectionRect()) {
-      ctx.save();
-      // Draw transformed preview only if we are NOT applying live preview to the layer
-      if (
-        (!this.freeTransformPreviewIndices ||
-          this.freeTransformPreviewIndices.length === 0) &&
-        this.freeTransformOriginalBuffer &&
-        this.freeTransformOriginalRect
-      ) {
-        const canvasW = this.document.canvasWidth();
-        const canvasH = this.document.canvasHeight();
-        const srcW = this.freeTransformOriginalRect.width;
-        const srcH = this.freeTransformOriginalRect.height;
-        const srcBuf = this.freeTransformOriginalBuffer;
-        const scaleX = ftState.width / srcW;
-        const scaleY = ftState.height / srcH;
-        const cosR = Math.cos((ftState.rotation * Math.PI) / 180);
-        const sinR = Math.sin((ftState.rotation * Math.PI) / 180);
-        const cX = srcW / 2;
-        const cY = srcH / 2;
-        for (let y = 0; y < ftState.height; y++) {
-          for (let x = 0; x < ftState.width; x++) {
-            const srcX = x / scaleX;
-            const srcY = y / scaleY;
-            const dx = srcX - cX;
-            const dy = srcY - cY;
-            const rotX = cX + dx * cosR + dy * sinR;
-            const rotY = cY - dx * sinR + dy * cosR;
-            const sx = Math.floor(rotX);
-            const sy = Math.floor(rotY);
-            if (sx >= 0 && sx < srcW && sy >= 0 && sy < srcH) {
-              const srcIdx = sy * srcW + sx;
-              const col = srcBuf[srcIdx];
-              if (col && col.length) {
-                const destX = Math.floor(ftState.x + x);
-                const destY = Math.floor(ftState.y + y);
-                if (
-                  destX >= 0 &&
-                  destX < canvasW &&
-                  destY >= 0 &&
-                  destY < canvasH
-                ) {
-                  ctx.fillStyle = col;
-                  ctx.fillRect(destX, destY, 1, 1);
-                }
-              }
-            }
-          }
-        }
-      }
-      // Draw rotated bounding box
-      ctx.lineWidth = pxLineWidth;
-      ctx.strokeStyle = isDark ? 'rgba(255,255,255,0.85)' : 'rgba(0,0,0,0.85)';
-      const rad = (ftState.rotation * Math.PI) / 180;
-      const cos = Math.cos(rad);
-      const sin = Math.sin(rad);
-      const cx = ftState.x + ftState.width / 2;
-      const cy = ftState.y + ftState.height / 2;
-      const corners = [
-        { dx: -ftState.width / 2, dy: -ftState.height / 2 },
-        { dx: ftState.width / 2, dy: -ftState.height / 2 },
-        { dx: ftState.width / 2, dy: ftState.height / 2 },
-        { dx: -ftState.width / 2, dy: ftState.height / 2 },
-      ].map((c) => ({
-        x: cx + c.dx * cos - c.dy * sin,
-        y: cy + c.dx * sin + c.dy * cos,
-      }));
-      ctx.beginPath();
-      ctx.moveTo(corners[0].x, corners[0].y);
-      for (let i = 1; i < corners.length; i++)
-        ctx.lineTo(corners[i].x, corners[i].y);
-      ctx.closePath();
-      ctx.stroke();
-
-      // Handles (smaller)
-      const handles: TransformHandle[] = [
-        'top-left',
-        'top-center',
-        'top-right',
-        'middle-left',
-        'middle-right',
-        'bottom-left',
-        'bottom-center',
-        'bottom-right',
-        'rotate-center',
-      ];
-      for (const h of handles) {
-        const pos = this.freeTransform.getHandlePosition(h, ftState);
-        const rOuter = Math.max(0.5, 0.8 / Math.max(0.001, scale));
-        ctx.fillStyle =
-          h === 'rotate-center'
-            ? isDark
-              ? '#3b82f6'
-              : '#1d4ed8'
-            : isDark
-              ? '#ffffff'
-              : '#000000';
-        ctx.beginPath();
-        ctx.arc(pos.x, pos.y, rOuter, 0, Math.PI * 2);
-        ctx.fill();
-        if (h === 'rotate-center') {
-          const rr = Math.max(2, 4 / Math.max(0.001, scale));
-          ctx.strokeStyle = ctx.fillStyle;
-          ctx.lineWidth = Math.max(pxLineWidth, 0.8 / Math.max(0.001, scale));
-          ctx.beginPath();
-          ctx.arc(pos.x, pos.y, rr, -Math.PI * 0.2, Math.PI * 1.1);
-          ctx.stroke();
-        }
-      }
-      const positions = this.computeTransformButtonPositions(ftState, scale);
-      const {
-        btnSize,
-        commitX,
-        commitY,
-        cancelX,
-        cancelY,
-        mirrorXX,
-        mirrorXY,
-        mirrorYX,
-        mirrorYY,
-      } = positions;
-      // mirror Y button (vertical arrows)
-      ctx.fillStyle = this.freeTransformMirrorY
-        ? isDark
-          ? '#047857'
-          : '#059669'
-        : isDark
-          ? '#065f46'
-          : '#047857';
-      ctx.fillRect(mirrorYX, mirrorYY, btnSize, btnSize);
-      ctx.strokeStyle = isDark ? '#d1d5db' : '#f9fafb';
-      ctx.lineWidth = Math.max(pxLineWidth, 0.8 / Math.max(0.001, scale));
-      ctx.beginPath();
-      ctx.moveTo(mirrorYX + btnSize * 0.5, mirrorYY + btnSize * 0.2);
-      ctx.lineTo(mirrorYX + btnSize * 0.5, mirrorYY + btnSize * 0.8);
-      ctx.moveTo(mirrorYX + btnSize * 0.5, mirrorYY + btnSize * 0.2);
-      ctx.lineTo(mirrorYX + btnSize * 0.35, mirrorYY + btnSize * 0.35);
-      ctx.moveTo(mirrorYX + btnSize * 0.5, mirrorYY + btnSize * 0.2);
-      ctx.lineTo(mirrorYX + btnSize * 0.65, mirrorYY + btnSize * 0.35);
-      ctx.moveTo(mirrorYX + btnSize * 0.5, mirrorYY + btnSize * 0.8);
-      ctx.lineTo(mirrorYX + btnSize * 0.35, mirrorYY + btnSize * 0.65);
-      ctx.moveTo(mirrorYX + btnSize * 0.5, mirrorYY + btnSize * 0.8);
-      ctx.lineTo(mirrorYX + btnSize * 0.65, mirrorYY + btnSize * 0.65);
-      ctx.stroke();
-
-      // mirror X button (horizontal arrows)
-      ctx.fillStyle = this.freeTransformMirrorX
-        ? isDark
-          ? '#047857'
-          : '#059669'
-        : isDark
-          ? '#065f46'
-          : '#047857';
-      ctx.fillRect(mirrorXX, mirrorXY, btnSize, btnSize);
-      ctx.beginPath();
-      ctx.moveTo(mirrorXX + btnSize * 0.2, mirrorXY + btnSize * 0.5);
-      ctx.lineTo(mirrorXX + btnSize * 0.8, mirrorXY + btnSize * 0.5);
-      ctx.moveTo(mirrorXX + btnSize * 0.2, mirrorXY + btnSize * 0.5);
-      ctx.lineTo(mirrorXX + btnSize * 0.35, mirrorXY + btnSize * 0.35);
-      ctx.moveTo(mirrorXX + btnSize * 0.2, mirrorXY + btnSize * 0.5);
-      ctx.lineTo(mirrorXX + btnSize * 0.35, mirrorXY + btnSize * 0.65);
-      ctx.moveTo(mirrorXX + btnSize * 0.8, mirrorXY + btnSize * 0.5);
-      ctx.lineTo(mirrorXX + btnSize * 0.65, mirrorXY + btnSize * 0.35);
-      ctx.moveTo(mirrorXX + btnSize * 0.8, mirrorXY + btnSize * 0.5);
-      ctx.lineTo(mirrorXX + btnSize * 0.65, mirrorXY + btnSize * 0.65);
-      ctx.stroke();
-
-      ctx.fillStyle = isDark ? '#1e40af' : '#1d4ed8';
-      ctx.fillRect(commitX, commitY, btnSize, btnSize);
-
-      ctx.strokeStyle = '#d1d5db';
-      ctx.lineWidth = Math.max(pxLineWidth, 0.8 / Math.max(0.001, scale));
-      ctx.beginPath();
-      ctx.moveTo(commitX + btnSize * 0.2, commitY + btnSize * 0.5);
-      ctx.lineTo(commitX + btnSize * 0.4, commitY + btnSize * 0.7);
-      ctx.lineTo(commitX + btnSize * 0.8, commitY + btnSize * 0.25);
-      ctx.stroke();
-
-      ctx.fillStyle = isDark ? '#991b1b' : '#b91c1c';
-      ctx.fillRect(cancelX, cancelY, btnSize, btnSize);
-      ctx.strokeStyle = '#d1d5db';
-      ctx.beginPath();
-      ctx.moveTo(cancelX + btnSize * 0.28, cancelY + btnSize * 0.28);
-      ctx.lineTo(cancelX + btnSize * 0.72, cancelY + btnSize * 0.72);
-      ctx.moveTo(cancelX + btnSize * 0.72, cancelY + btnSize * 0.28);
-      ctx.lineTo(cancelX + btnSize * 0.28, cancelY + btnSize * 0.72);
-      ctx.stroke();
-      ctx.restore();
-    }
-
-    const distortState = this.distort.distortState();
-    if (distortState) {
-      ctx.save();
-      ctx.lineWidth = pxLineWidth;
-      ctx.strokeStyle = isDark ? 'rgba(59,130,246,0.9)' : 'rgba(37,99,235,0.9)';
-
-      const corners = distortState.corners;
-      ctx.beginPath();
-      ctx.moveTo(corners.topLeft.x, corners.topLeft.y);
-      ctx.lineTo(corners.topRight.x, corners.topRight.y);
-      ctx.lineTo(corners.bottomRight.x, corners.bottomRight.y);
-      ctx.lineTo(corners.bottomLeft.x, corners.bottomLeft.y);
-      ctx.closePath();
-      ctx.stroke();
-
-      ctx.strokeStyle = isDark ? 'rgba(59,130,246,0.5)' : 'rgba(37,99,235,0.5)';
-      const gridSteps = 3;
-      for (let i = 1; i < gridSteps; i++) {
-        const t = i / gridSteps;
-        const topX =
-          corners.topLeft.x + (corners.topRight.x - corners.topLeft.x) * t;
-        const topY =
-          corners.topLeft.y + (corners.topRight.y - corners.topLeft.y) * t;
-        const bottomX =
-          corners.bottomLeft.x +
-          (corners.bottomRight.x - corners.bottomLeft.x) * t;
-        const bottomY =
-          corners.bottomLeft.y +
-          (corners.bottomRight.y - corners.bottomLeft.y) * t;
-        ctx.beginPath();
-        ctx.moveTo(topX, topY);
-        ctx.lineTo(bottomX, bottomY);
-        ctx.stroke();
-
-        const leftX =
-          corners.topLeft.x + (corners.bottomLeft.x - corners.topLeft.x) * t;
-        const leftY =
-          corners.topLeft.y + (corners.bottomLeft.y - corners.topLeft.y) * t;
-        const rightX =
-          corners.topRight.x + (corners.bottomRight.x - corners.topRight.x) * t;
-        const rightY =
-          corners.topRight.y + (corners.bottomRight.y - corners.topRight.y) * t;
-        ctx.beginPath();
-        ctx.moveTo(leftX, leftY);
-        ctx.lineTo(rightX, rightY);
-        ctx.stroke();
-      }
-
-      const handles: DistortHandle[] = [
-        'top-left',
-        'top-right',
-        'bottom-right',
-        'bottom-left',
-      ];
-      for (const h of handles) {
-        const pos = this.distort.getHandlePosition(h);
-        if (pos) {
-          const rOuter = Math.max(1.2, 1.5 / Math.max(0.001, scale));
-          ctx.fillStyle = isDark ? '#3b82f6' : '#2563eb';
-          ctx.beginPath();
-          ctx.arc(pos.x, pos.y, rOuter, 0, Math.PI * 2);
-          ctx.fill();
-          ctx.strokeStyle = isDark ? '#ffffff' : '#000000';
-          ctx.lineWidth = Math.max(pxLineWidth, 0.5 / Math.max(0.001, scale));
-          ctx.beginPath();
-          ctx.arc(pos.x, pos.y, rOuter, 0, Math.PI * 2);
-          ctx.stroke();
-        }
-      }
-
-      const buttonPositions = this.computeDistortButtonPositions(
-        distortState,
-        scale,
-      );
-      const { btnSize, commitX, commitY, cancelX, cancelY } = buttonPositions;
-
-      ctx.fillStyle = isDark ? '#059669' : '#047857';
-      ctx.fillRect(commitX, commitY, btnSize, btnSize);
-      ctx.strokeStyle = isDark ? '#d1d5db' : '#f9fafb';
-      ctx.lineWidth = Math.max(pxLineWidth, 0.8 / Math.max(0.001, scale));
-      ctx.beginPath();
-      ctx.moveTo(commitX + btnSize * 0.2, commitY + btnSize * 0.5);
-      ctx.lineTo(commitX + btnSize * 0.4, commitY + btnSize * 0.7);
-      ctx.lineTo(commitX + btnSize * 0.8, commitY + btnSize * 0.25);
-      ctx.stroke();
-
-      ctx.fillStyle = isDark ? '#dc2626' : '#991b1b';
-      ctx.fillRect(cancelX, cancelY, btnSize, btnSize);
-      ctx.beginPath();
-      ctx.moveTo(cancelX + btnSize * 0.28, cancelY + btnSize * 0.28);
-      ctx.lineTo(cancelX + btnSize * 0.72, cancelY + btnSize * 0.72);
-      ctx.moveTo(cancelX + btnSize * 0.72, cancelY + btnSize * 0.28);
-      ctx.lineTo(cancelX + btnSize * 0.28, cancelY + btnSize * 0.72);
-      ctx.stroke();
-
-      ctx.restore();
-    }
-
-    const perspectiveState = this.perspective.perspectiveState();
-    if (perspectiveState) {
-      ctx.save();
-      ctx.lineWidth = pxLineWidth;
-      ctx.setLineDash([4 / Math.max(0.001, scale), 4 / Math.max(0.001, scale)]);
-      ctx.strokeStyle = isDark
-        ? 'rgba(147,51,234,0.9)'
-        : 'rgba(126,34,206,0.9)';
-
-      const corners = perspectiveState.corners;
-      ctx.beginPath();
-      ctx.moveTo(corners.topLeft.x, corners.topLeft.y);
-      ctx.lineTo(corners.topRight.x, corners.topRight.y);
-      ctx.lineTo(corners.bottomRight.x, corners.bottomRight.y);
-      ctx.lineTo(corners.bottomLeft.x, corners.bottomLeft.y);
-      ctx.closePath();
-      ctx.stroke();
-
-      ctx.strokeStyle = isDark
-        ? 'rgba(147,51,234,0.5)'
-        : 'rgba(126,34,206,0.5)';
-      const gridSteps = 3;
-      for (let i = 1; i < gridSteps; i++) {
-        const t = i / gridSteps;
-        const topX =
-          corners.topLeft.x + (corners.topRight.x - corners.topLeft.x) * t;
-        const topY =
-          corners.topLeft.y + (corners.topRight.y - corners.topLeft.y) * t;
-        const bottomX =
-          corners.bottomLeft.x +
-          (corners.bottomRight.x - corners.bottomLeft.x) * t;
-        const bottomY =
-          corners.bottomLeft.y +
-          (corners.bottomRight.y - corners.bottomLeft.y) * t;
-        ctx.beginPath();
-        ctx.moveTo(topX, topY);
-        ctx.lineTo(bottomX, bottomY);
-        ctx.stroke();
-
-        const leftX =
-          corners.topLeft.x + (corners.bottomLeft.x - corners.topLeft.x) * t;
-        const leftY =
-          corners.topLeft.y + (corners.bottomLeft.y - corners.topLeft.y) * t;
-        const rightX =
-          corners.topRight.x + (corners.bottomRight.x - corners.topRight.x) * t;
-        const rightY =
-          corners.topRight.y + (corners.bottomRight.y - corners.topRight.y) * t;
-        ctx.beginPath();
-        ctx.moveTo(leftX, leftY);
-        ctx.lineTo(rightX, rightY);
-        ctx.stroke();
-      }
-
-      ctx.setLineDash([]);
-
-      const handles: PerspectiveHandle[] = [
-        'top-left',
-        'top-right',
-        'bottom-right',
-        'bottom-left',
-      ];
-      for (const h of handles) {
-        const pos = this.perspective.getHandlePosition(h);
-        if (pos) {
-          const rOuter = Math.max(1.2, 1.5 / Math.max(0.001, scale));
-          ctx.fillStyle = isDark ? '#9333ea' : '#7e22ce';
-          ctx.beginPath();
-          ctx.arc(pos.x, pos.y, rOuter, 0, Math.PI * 2);
-          ctx.fill();
-          ctx.strokeStyle = isDark ? '#ffffff' : '#000000';
-          ctx.lineWidth = Math.max(pxLineWidth, 0.5 / Math.max(0.001, scale));
-          ctx.beginPath();
-          ctx.arc(pos.x, pos.y, rOuter, 0, Math.PI * 2);
-          ctx.stroke();
-        }
-      }
-
-      const buttonPositions = this.computePerspectiveButtonPositions(
-        perspectiveState,
-        scale,
-      );
-      const { btnSize, commitX, commitY, cancelX, cancelY } = buttonPositions;
-
-      ctx.fillStyle = isDark ? '#059669' : '#047857';
-      ctx.fillRect(commitX, commitY, btnSize, btnSize);
-      ctx.strokeStyle = isDark ? '#d1d5db' : '#f9fafb';
-      ctx.lineWidth = Math.max(pxLineWidth, 0.8 / Math.max(0.001, scale));
-      ctx.beginPath();
-      ctx.moveTo(commitX + btnSize * 0.2, commitY + btnSize * 0.5);
-      ctx.lineTo(commitX + btnSize * 0.4, commitY + btnSize * 0.7);
-      ctx.lineTo(commitX + btnSize * 0.8, commitY + btnSize * 0.25);
-      ctx.stroke();
-
-      ctx.fillStyle = isDark ? '#dc2626' : '#991b1b';
-      ctx.fillRect(cancelX, cancelY, btnSize, btnSize);
-      ctx.beginPath();
-      ctx.moveTo(cancelX + btnSize * 0.28, cancelY + btnSize * 0.28);
-      ctx.lineTo(cancelX + btnSize * 0.72, cancelY + btnSize * 0.72);
-      ctx.moveTo(cancelX + btnSize * 0.72, cancelY + btnSize * 0.28);
-      ctx.lineTo(cancelX + btnSize * 0.28, cancelY + btnSize * 0.72);
-      ctx.stroke();
-
-      ctx.restore();
-    }
-
-    const warpState = this.warp.warpState();
-    if (warpState) {
-      ctx.save();
-      ctx.lineWidth = pxLineWidth;
-      ctx.strokeStyle = isDark ? 'rgba(34,197,94,0.9)' : 'rgba(22,163,74,0.9)';
-
-      const dims = this.warp.getGridDimensions();
-      if (dims) {
-        const { rows, cols } = dims;
-
-        for (let row = 0; row <= rows; row++) {
-          for (let col = 0; col <= cols; col++) {
-            const node = this.warp.getNode(row, col);
-            if (!node) continue;
-
-            if (col < cols) {
-              const nextNode = this.warp.getNode(row, col + 1);
-              if (nextNode) {
-                ctx.beginPath();
-                ctx.moveTo(node.x, node.y);
-                ctx.lineTo(nextNode.x, nextNode.y);
-                ctx.stroke();
-              }
-            }
-
-            if (row < rows) {
-              const nextNode = this.warp.getNode(row + 1, col);
-              if (nextNode) {
-                ctx.beginPath();
-                ctx.moveTo(node.x, node.y);
-                ctx.lineTo(nextNode.x, nextNode.y);
-                ctx.stroke();
-              }
-            }
-          }
-        }
-
-        for (const node of warpState.nodes) {
-          const rOuter = Math.max(1.2, 1.5 / Math.max(0.001, scale));
-          ctx.fillStyle = isDark ? '#22c55e' : '#16a34a';
-          ctx.beginPath();
-          ctx.arc(node.x, node.y, rOuter, 0, Math.PI * 2);
-          ctx.fill();
-          ctx.strokeStyle = isDark ? '#ffffff' : '#000000';
-          ctx.lineWidth = Math.max(pxLineWidth, 0.5 / Math.max(0.001, scale));
-          ctx.beginPath();
-          ctx.arc(node.x, node.y, rOuter, 0, Math.PI * 2);
-          ctx.stroke();
-        }
-      }
-
-      const buttonPositions = this.computeWarpButtonPositions(warpState, scale);
-      const { btnSize, commitX, commitY, cancelX, cancelY } = buttonPositions;
-
-      ctx.fillStyle = isDark ? '#059669' : '#047857';
-      ctx.fillRect(commitX, commitY, btnSize, btnSize);
-      ctx.strokeStyle = isDark ? '#d1d5db' : '#f9fafb';
-      ctx.lineWidth = Math.max(pxLineWidth, 0.8 / Math.max(0.001, scale));
-      ctx.beginPath();
-      ctx.moveTo(commitX + btnSize * 0.2, commitY + btnSize * 0.5);
-      ctx.lineTo(commitX + btnSize * 0.4, commitY + btnSize * 0.7);
-      ctx.lineTo(commitX + btnSize * 0.8, commitY + btnSize * 0.25);
-      ctx.stroke();
-
-      ctx.fillStyle = isDark ? '#dc2626' : '#991b1b';
-      ctx.fillRect(cancelX, cancelY, btnSize, btnSize);
-      ctx.beginPath();
-      ctx.moveTo(cancelX + btnSize * 0.28, cancelY + btnSize * 0.28);
-      ctx.lineTo(cancelX + btnSize * 0.72, cancelY + btnSize * 0.72);
-      ctx.moveTo(cancelX + btnSize * 0.72, cancelY + btnSize * 0.28);
-      ctx.lineTo(cancelX + btnSize * 0.28, cancelY + btnSize * 0.72);
-      ctx.stroke();
-
-      ctx.restore();
-    }
-
-    const puppetWarpState = this.puppetWarp.puppetWarpState();
-    if (puppetWarpState) {
-      ctx.save();
-
-      for (const pin of puppetWarpState.pins) {
-        ctx.fillStyle = isDark
-          ? 'rgba(236,72,153,0.2)'
-          : 'rgba(219,39,119,0.2)';
-        ctx.beginPath();
-        ctx.arc(pin.x, pin.y, pin.radius, 0, Math.PI * 2);
-        ctx.fill();
-        ctx.strokeStyle = isDark
-          ? 'rgba(236,72,153,0.4)'
-          : 'rgba(219,39,119,0.4)';
-        ctx.lineWidth = pxLineWidth;
-        ctx.setLineDash([
-          4 / Math.max(0.001, scale),
-          4 / Math.max(0.001, scale),
-        ]);
-        ctx.stroke();
-        ctx.setLineDash([]);
-
-        const rOuter = Math.max(1.5, 2 / Math.max(0.001, scale));
-        ctx.fillStyle = pin.locked
-          ? isDark
-            ? '#ef4444'
-            : '#dc2626'
-          : isDark
-            ? '#ec4899'
-            : '#db2777';
-        ctx.beginPath();
-        ctx.arc(pin.x, pin.y, rOuter, 0, Math.PI * 2);
-        ctx.fill();
-        ctx.strokeStyle = isDark ? '#ffffff' : '#000000';
-        ctx.lineWidth = Math.max(pxLineWidth, 0.5 / Math.max(0.001, scale));
-        ctx.beginPath();
-        ctx.arc(pin.x, pin.y, rOuter, 0, Math.PI * 2);
-        ctx.stroke();
-
-        if (pin.locked) {
-          const lockSize = rOuter * 0.8;
-          ctx.fillStyle = isDark ? '#ffffff' : '#000000';
-          ctx.fillRect(
-            pin.x - lockSize * 0.3,
-            pin.y - lockSize * 0.5,
-            lockSize * 0.6,
-            lockSize,
-          );
-        }
-      }
-
-      const buttonPositions = this.computePuppetWarpButtonPositions(
-        puppetWarpState,
-        scale,
-      );
-      const { btnSize, commitX, commitY, cancelX, cancelY } = buttonPositions;
-
-      ctx.fillStyle = isDark ? '#059669' : '#047857';
-      ctx.fillRect(commitX, commitY, btnSize, btnSize);
-      ctx.strokeStyle = isDark ? '#d1d5db' : '#f9fafb';
-      ctx.lineWidth = Math.max(pxLineWidth, 0.8 / Math.max(0.001, scale));
-      ctx.beginPath();
-      ctx.moveTo(commitX + btnSize * 0.2, commitY + btnSize * 0.5);
-      ctx.lineTo(commitX + btnSize * 0.4, commitY + btnSize * 0.7);
-      ctx.lineTo(commitX + btnSize * 0.8, commitY + btnSize * 0.25);
-      ctx.stroke();
-
-      ctx.fillStyle = isDark ? '#dc2626' : '#991b1b';
-      ctx.fillRect(cancelX, cancelY, btnSize, btnSize);
-      ctx.beginPath();
-      ctx.moveTo(cancelX + btnSize * 0.28, cancelY + btnSize * 0.28);
-      ctx.lineTo(cancelX + btnSize * 0.72, cancelY + btnSize * 0.72);
-      ctx.moveTo(cancelX + btnSize * 0.72, cancelY + btnSize * 0.28);
-      ctx.lineTo(cancelX + btnSize * 0.28, cancelY + btnSize * 0.72);
-      ctx.stroke();
-
-      ctx.restore();
-    }
-  }
-
-  private computeTransformButtonPositions(
-    ftState: { x: number; y: number; width: number; height: number },
-    scale: number,
-  ): {
-    btnSize: number;
-    commitX: number;
-    commitY: number;
-    cancelX: number;
-    cancelY: number;
-    mirrorXX: number;
-    mirrorXY: number;
-    mirrorYX: number;
-    mirrorYY: number;
-  } {
-    return this.renderService.computeTransformButtonPositions(
-      ftState,
-      scale,
-      this.document.canvasWidth(),
-      this.document.canvasHeight(),
-    );
-  }
-
-  private computeDistortButtonPositions(
-    distortState: {
-      corners: any;
-      sourceX: number;
-      sourceY: number;
-      sourceWidth: number;
-      sourceHeight: number;
-    },
-    scale: number,
-  ): {
-    btnSize: number;
-    commitX: number;
-    commitY: number;
-    cancelX: number;
-    cancelY: number;
-  } {
-    return this.renderService.computeDistortButtonPositions(
-      distortState,
-      scale,
-      this.document.canvasWidth(),
-      this.document.canvasHeight(),
-    );
-  }
-
-  private computePerspectiveButtonPositions(
-    perspectiveState: {
-      corners: any;
-      sourceX: number;
-      sourceY: number;
-      sourceWidth: number;
-      sourceHeight: number;
-    },
-    scale: number,
-  ): {
-    btnSize: number;
-    commitX: number;
-    commitY: number;
-    cancelX: number;
-    cancelY: number;
-  } {
-    return this.renderService.computeDistortButtonPositions(
-      perspectiveState,
-      scale,
-      this.document.canvasWidth(),
-      this.document.canvasHeight(),
-    );
-  }
-
-  private computeWarpButtonPositions(
-    warpState: {
-      sourceX: number;
-      sourceY: number;
-      sourceWidth: number;
-      sourceHeight: number;
-      nodes: any[];
-    },
-    scale: number,
-  ): {
-    btnSize: number;
-    commitX: number;
-    commitY: number;
-    cancelX: number;
-    cancelY: number;
-  } {
-    return this.renderService.computeWarpButtonPositions(
-      warpState,
-      scale,
-      this.document.canvasWidth(),
-      this.document.canvasHeight(),
-    );
-  }
-
-  private computePuppetWarpButtonPositions(
-    puppetWarpState: {
-      sourceX: number;
-      sourceY: number;
-      sourceWidth: number;
-      sourceHeight: number;
-      pins: any[];
-    },
-    scale: number,
-  ): {
-    btnSize: number;
-    commitX: number;
-    commitY: number;
-    cancelX: number;
-    cancelY: number;
-  } {
-    return this.renderService.computePuppetWarpButtonPositions(
-      puppetWarpState,
-      scale,
-      this.document.canvasWidth(),
-      this.document.canvasHeight(),
-    );
+    this.drawService.drawCanvas({
+      canvas,
+      container: this.canvasContainer?.nativeElement,
+      workspaceWidth: this.workspaceWidth(),
+      workspaceHeight: this.workspaceHeight(),
+      tileSize: this.tileSize(),
+      mouseX: this.mouseX(),
+      mouseY: this.mouseY(),
+      activeShapeTool: this.activeShapeTool(),
+      shapeStart: this.shapeStart(),
+      shapeCurrent: this.shapeCurrent(),
+      shapeConstrainUniform: this.shapeConstrainUniform(),
+      penDrawing: this.penDrawing(),
+      penPoints: this.penPoints(),
+      movingContentBuffer: this.movingContentBuffer,
+      movingContentOriginalRect: this.movingContentOriginalRect,
+      getCurrentFrameId: () => this.getCurrentFrameId(),
+      getCircleDrawOptions: () => this.getCircleDrawOptions(),
+      getSquareDrawOptions: () => this.getSquareDrawOptions(),
+    });
   }
 
   private startShape(mode: 'line' | 'circle' | 'square', x: number, y: number) {
     const width = Math.max(1, this.document.canvasWidth());
     const height = Math.max(1, this.document.canvasHeight());
-    const point = {
-      x: this.clampCoord(x, width),
-      y: this.clampCoord(y, height),
-    };
+    const point = { x: this.clampCoord(x, width), y: this.clampCoord(y, height) };
     this.document.saveSnapshot(mode);
-    this.shaping = true;
+    this.pointerState.shaping = true;
     this.activeShapeTool.set(mode);
     this.shapeStart.set(point);
     this.shapeCurrent.set(point);
   }
 
   private finishShape(constrainOverride?: boolean) {
-    if (!this.shaping) return;
+    if (!this.pointerState.shaping) return;
     const mode = this.activeShapeTool();
     const start = this.shapeStart();
     const current = this.shapeCurrent();
-    if (!mode || !start || !current) {
-      this.clearShapeState();
-      return;
-    }
+    if (!mode || !start || !current) { this.clearShapeState(); return; }
     const layerId = this.document.selectedLayerId();
-    if (!layerId) {
-      this.clearShapeState();
-      return;
-    }
+    if (!layerId) { this.clearShapeState(); return; }
     if (mode === 'line') {
       const thickness = this.tools.lineThickness();
       const color = this.tools.lineColor();
-      this.document.applyLineToLayer(
-        layerId,
-        start.x,
-        start.y,
-        current.x,
-        current.y,
-        color,
-        thickness,
-      );
+      this.document.applyLineToLayer(layerId, start.x, start.y, current.x, current.y, color, thickness);
     } else if (mode === 'circle') {
-      this.document.applyCircleToLayer(
-        layerId,
-        start.x,
-        start.y,
-        current.x,
-        current.y,
-        this.getCircleDrawOptions(),
-        typeof constrainOverride === 'boolean'
-          ? constrainOverride
-          : this.shapeConstrainUniform(),
-      );
+      this.document.applyCircleToLayer(layerId, start.x, start.y, current.x, current.y, this.getCircleDrawOptions(),
+        typeof constrainOverride === 'boolean' ? constrainOverride : this.shapeConstrainUniform());
     } else {
-      const constrainSquare =
-        typeof constrainOverride === 'boolean'
-          ? constrainOverride
-          : this.shapeConstrainUniform();
-      this.document.applySquareToLayer(
-        layerId,
-        start.x,
-        start.y,
-        current.x,
-        current.y,
-        this.getSquareDrawOptions(),
-        constrainSquare,
-      );
+      const constrainSquare = typeof constrainOverride === 'boolean' ? constrainOverride : this.shapeConstrainUniform();
+      this.document.applySquareToLayer(layerId, start.x, start.y, current.x, current.y, this.getSquareDrawOptions(), constrainSquare);
     }
-    this.clearShapeState();
-  }
-
-  private cancelShape() {
-    if (!this.shaping) return;
     this.clearShapeState();
   }
 
   private clearShapeState() {
-    this.shaping = false;
+    this.pointerState.shaping = false;
     this.activeShapeTool.set(null);
     this.shapeStart.set(null);
     this.shapeCurrent.set(null);
@@ -4791,9 +924,7 @@ export class EditorCanvas implements OnDestroy {
 
   onDoubleClick(ev: MouseEvent) {
     const tool = this.tools.currentTool();
-    if (tool === 'pen' && this.penDrawing()) {
-      this.finishPenPath();
-    }
+    if (tool === 'pen' && this.penDrawing()) this.finishPenPath();
   }
 
   private addPenPoint(x: number, y: number) {
@@ -4802,100 +933,49 @@ export class EditorCanvas implements OnDestroy {
     const clampedX = Math.max(0, Math.min(w - 1, x));
     const clampedY = Math.max(0, Math.min(h - 1, y));
     const points = this.penPoints();
-    if (!this.penDrawing()) {
-      this.document.saveSnapshot('Pen path');
-      this.penDrawing.set(true);
-    }
+    if (!this.penDrawing()) { this.document.saveSnapshot('Pen path'); this.penDrawing.set(true); }
     this.penPoints.set([...points, { x: clampedX, y: clampedY }]);
   }
 
   private finishPenPath() {
     const points = this.penPoints();
-    if (points.length < 2) {
-      this.cancelPenPath();
-      return;
-    }
+    if (points.length < 2) { this.cancelPenPath(); return; }
     const layerId = this.document.selectedLayerId();
-    if (!layerId) {
-      this.cancelPenPath();
-      return;
-    }
+    if (!layerId) { this.cancelPenPath(); return; }
     const thickness = this.tools.penThickness();
     const color = this.tools.penColor();
     const lineMode = this.tools.penLineMode();
-    if (lineMode === 'spline') {
-      this.applySplinePath(layerId, points, color, thickness);
-    } else {
-      this.applyPolylinePath(layerId, points, color, thickness);
-    }
+    if (lineMode === 'spline') this.applySplinePath(layerId, points, color, thickness);
+    else this.applyPolylinePath(layerId, points, color, thickness);
     this.clearPenState();
   }
 
   private cancelPenPath() {
-    if (this.penDrawing()) {
-      this.document.undo();
-    }
+    if (this.penDrawing()) this.document.undo();
     this.clearPenState();
   }
 
-  private clearPenState() {
-    this.penDrawing.set(false);
-    this.penPoints.set([]);
-  }
+  private clearPenState() { this.penDrawing.set(false); this.penPoints.set([]); }
 
-  private applyPolylinePath(
-    layerId: string,
-    points: { x: number; y: number }[],
-    color: string,
-    thickness: number,
-  ) {
+  private applyPolylinePath(layerId: string, points: { x: number; y: number }[], color: string, thickness: number) {
     for (let i = 0; i < points.length - 1; i++) {
       const start = points[i];
       const end = points[i + 1];
-      this.document.applyLineToLayer(
-        layerId,
-        start.x,
-        start.y,
-        end.x,
-        end.y,
-        color,
-        thickness,
-      );
+      this.document.applyLineToLayer(layerId, start.x, start.y, end.x, end.y, color, thickness);
     }
   }
 
-  private applySplinePath(
-    layerId: string,
-    points: { x: number; y: number }[],
-    color: string,
-    thickness: number,
-  ) {
+  private applySplinePath(layerId: string, points: { x: number; y: number }[], color: string, thickness: number) {
     if (points.length < 2) return;
     if (points.length === 2) {
-      this.document.applyLineToLayer(
-        layerId,
-        points[0].x,
-        points[0].y,
-        points[1].x,
-        points[1].y,
-        color,
-        thickness,
-      );
+      this.document.applyLineToLayer(layerId, points[0].x, points[0].y, points[1].x, points[1].y, color, thickness);
       return;
     }
     const splinePoints = this.shapeService.catmullRomSpline(points, 10);
     for (let i = 0; i < splinePoints.length - 1; i++) {
       const start = splinePoints[i];
       const end = splinePoints[i + 1];
-      this.document.applyLineToLayer(
-        layerId,
-        Math.round(start.x),
-        Math.round(start.y),
-        Math.round(end.x),
-        Math.round(end.y),
-        color,
-        thickness,
-      );
+      this.document.applyLineToLayer(layerId, Math.round(start.x), Math.round(start.y), Math.round(end.x), Math.round(end.y), color, thickness);
     }
   }
 
@@ -4920,71 +1000,30 @@ export class EditorCanvas implements OnDestroy {
       const cy = rect.y + (rect.height - 1) / 2;
       const normX = x - cx;
       const normY = y - cy;
-      const ellipseTest =
-        (normX * normX) / (rx * rx) + (normY * normY) / (ry * ry);
+      const ellipseTest = (normX * normX) / (rx * rx) + (normY * normY) / (ry * ry);
       return ellipseTest <= 1;
     }
     return withinRect;
   }
 
-  private pointInPolygon(
-    x: number,
-    y: number,
-    polygon: { x: number; y: number }[],
-  ): boolean {
+  private pointInPolygon(x: number, y: number, polygon: { x: number; y: number }[]): boolean {
     let inside = false;
     for (let i = 0, j = polygon.length - 1; i < polygon.length; j = i++) {
       const xi = polygon[i].x;
       const yi = polygon[i].y;
       const xj = polygon[j].x;
       const yj = polygon[j].y;
-      const intersects =
-        yi > y !== yj > y &&
-        x <= ((xj - xi) * (y - yi)) / (yj - yi || Number.EPSILON) + xi;
+      const intersects = yi > y !== yj > y && x <= ((xj - xi) * (y - yi)) / (yj - yi || Number.EPSILON) + xi;
       if (intersects) inside = !inside;
     }
     return inside;
   }
 
-  private clampCoord(value: number, max: number) {
-    return Math.max(0, Math.min(Math.floor(value), max - 1));
-  }
-
-  private computeRectBounds(
-    start: { x: number; y: number },
-    current: { x: number; y: number },
-    constrainToSquare: boolean,
-  ) {
-    const width = Math.max(1, this.document.canvasWidth());
-    const height = Math.max(1, this.document.canvasHeight());
-    const sx = this.clampCoord(start.x, width);
-    const sy = this.clampCoord(start.y, height);
-    const cx = this.clampCoord(current.x, width);
-    const cy = this.clampCoord(current.y, height);
-    const dx = cx - sx;
-    const dy = cy - sy;
-    let ex = cx;
-    let ey = cy;
-    if (constrainToSquare) {
-      const stepX = dx >= 0 ? 1 : -1;
-      const stepY = dy >= 0 ? 1 : -1;
-      const span = Math.max(Math.abs(dx), Math.abs(dy));
-      ex = this.clampCoord(sx + stepX * span, width);
-      ey = this.clampCoord(sy + stepY * span, height);
-    }
-    const minX = Math.max(0, Math.min(sx, ex));
-    const maxX = Math.min(width - 1, Math.max(sx, ex));
-    const minY = Math.max(0, Math.min(sy, ey));
-    const maxY = Math.min(height - 1, Math.max(sy, ey));
-    return { minX, minY, maxX, maxY };
-  }
+  private clampCoord(value: number, max: number) { return Math.max(0, Math.min(Math.floor(value), max - 1)); }
 
   private getCircleDrawOptions(): ShapeDrawOptions {
     return {
-      strokeThickness: Math.max(
-        0,
-        Math.floor(this.tools.circleStrokeThickness()),
-      ),
+      strokeThickness: Math.max(0, Math.floor(this.tools.circleStrokeThickness())),
       strokeColor: this.tools.circleStrokeColor(),
       fillMode: this.tools.circleFillMode(),
       fillColor: this.tools.circleFillColor(),
@@ -4997,10 +1036,7 @@ export class EditorCanvas implements OnDestroy {
 
   private getSquareDrawOptions(): ShapeDrawOptions {
     return {
-      strokeThickness: Math.max(
-        0,
-        Math.floor(this.tools.squareStrokeThickness()),
-      ),
+      strokeThickness: Math.max(0, Math.floor(this.tools.squareStrokeThickness())),
       strokeColor: this.tools.squareStrokeColor(),
       fillMode: this.tools.squareFillMode(),
       fillColor: this.tools.squareFillColor(),
@@ -5009,24 +1045,6 @@ export class EditorCanvas implements OnDestroy {
       gradientType: this.tools.squareGradientType(),
       gradientAngle: this.tools.squareGradientAngle(),
     };
-  }
-
-  private renderSquarePreview(
-    ctx: CanvasRenderingContext2D,
-    bounds: { minX: number; minY: number; maxX: number; maxY: number },
-    options: ShapeDrawOptions,
-    pxLineWidth: number,
-  ) {
-    this.renderService.renderSquarePreview(ctx, bounds, options, pxLineWidth);
-  }
-
-  private renderEllipsePreview(
-    ctx: CanvasRenderingContext2D,
-    bounds: { minX: number; minY: number; maxX: number; maxY: number },
-    options: ShapeDrawOptions,
-    pxLineWidth: number,
-  ) {
-    this.renderService.renderEllipsePreview(ctx, bounds, options, pxLineWidth);
   }
 
   private measureContainer() {
@@ -5041,48 +1059,16 @@ export class EditorCanvas implements OnDestroy {
     return frame?.id || '';
   }
 
-  private findBonePointAt(
-    frameId: string,
-    x: number,
-    y: number,
-  ): { boneId: string; pointId: string } | null {
-    const bones = this.boneService.getBones(frameId);
-    const hitRadius = 5;
-
-    for (const bone of bones) {
-      for (const point of bone.points) {
-        const dx = point.x - x;
-        const dy = point.y - y;
-        const dist = Math.sqrt(dx * dx + dy * dy);
-        if (dist <= hitRadius) {
-          return { boneId: bone.id, pointId: point.id };
-        }
-      }
-    }
-    return null;
-  }
-
   private handleGenerateFromSelection(sourceType: 'layer' | 'visible'): void {
     const sel = this.document.selectionRect();
-    if (!sel || sel.width <= 0 || sel.height <= 0) {
-      return;
-    }
-
+    if (!sel || sel.width <= 0 || sel.height <= 0) return;
     const sketchDataUrl = this.generationService.extractSketchFromSelection(
       sourceType,
       sel,
       (px, py) => this.isPointInSelection(px, py),
     );
-    if (!sketchDataUrl) {
-      return;
-    }
-
-    this.pixelGenerationDialog?.show(
-      sketchDataUrl,
-      sel.width,
-      sel.height,
-      'selection',
-    );
+    if (!sketchDataUrl) return;
+    this.pixelGenerationDialog?.show(sketchDataUrl, sel.width, sel.height, 'selection');
   }
 
   handleGenerate(request: GeneratePixelArtRequest): void {
