+++ src/app/services/editor/canvas/index.ts
@@ -4,3 +4,11 @@ export * from './canvas-selection-move.service';
 export * from './canvas-transform-state.service';
 export * from './canvas-render.service';
 export * from './canvas-generation.service';
+export * from './canvas-free-transform.handler';
+export * from './canvas-distort.handler';
+export * from './canvas-perspective.handler';
+export * from './canvas-warp.handler';
+export * from './canvas-puppet-warp.handler';
+export * from './canvas-context-menu.service';
+export * from './canvas-draw.service';
+export * from './canvas-pointer.service';
